refactor(coordinator/store): Phase 2 — split memory.go & postgres.go into per-domain files + dedup
Both monolith impls are split along the same per-domain boundaries as the Phase 1
interface fragments, mirroring each other file-for-file:
  postgres.go  3561 -> 663 lines  (+ 17 postgres_*.go files)
  memory.go    2641 -> 211 lines  (+ 17 memory_*.go files)
Each split is a strict PURE MOVE (verified deletion-only on the source files; no
method body, SQL string, or comment changed). manifestFromRecord moved to a
shared model_registry_common.go (used by both backends).

Dedup collapses (logic-touching, verified):
- providerColumns const + scanProviderRow(rowScanner): the 26-column provider
  SELECT/Scan, previously copy-pasted across 4 read methods, now lives once.
- NormalizeLeaderboardLimit: the <=0||>200 -> 50 clamp shared by both backends.

Verified: gofmt clean, go build/vet ./coordinator/..., and
go test ./coordinator/store/... -race against a throwaway Postgres 16
(DATABASE_URL set) — all green. (Pre-existing flaky integration test
TestPostgresRecordUsage, which asserts insertion order against an
ORDER BY created_at DESC query, is unrelated and excluded; it fails
identically on master.)
