+++ coordinator/store/postgres.go
@@ -15,14 +15,10 @@ package store
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"strconv"
 	"sync"
 	"time"
 
-	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
@@ -664,2898 +660,4 @@ func (s *PostgresStore) migrate(ctx context.Context) error {
 		}
 	}
 	return nil
-}
-
-// apiKeyColumns is the canonical SELECT list for reading an api_keys row into
-// an APIKey via scanAPIKeyRow.
-const apiKeyColumns = `id, owner_account_id, name, raw_prefix, key_hash, active,
-	limit_micro_usd, limit_reset, rpm_limit, itpm_limit, otpm_limit,
-	allowed_models, expires_at, created_at, last_used_at`
-
-// rowScanner is satisfied by both pgx.Row and pgx.Rows.
-type rowScanner interface {
-	Scan(dest ...any) error
-}
-
-// scanAPIKeyRow scans one api_keys row (selected via apiKeyColumns) into APIKey.
-func scanAPIKeyRow(row rowScanner) (*APIKey, error) {
-	var (
-		k          APIKey
-		active     bool
-		limit      *int64
-		rpm        *int64
-		itpm       *int64
-		otpm       *int64
-		allowed    string
-		expiresAt  *time.Time
-		lastUsedAt *time.Time
-	)
-	if err := row.Scan(&k.ID, &k.OwnerAccountID, &k.Name, &k.Label, &k.KeyHash, &active,
-		&limit, &k.LimitReset, &rpm, &itpm, &otpm,
-		&allowed, &expiresAt, &k.CreatedAt, &lastUsedAt); err != nil {
-		return nil, err
-	}
-	k.Disabled = !active
-	k.LimitMicroUSD = limit
-	k.RPMLimit = rpm
-	k.ITPMLimit = itpm
-	k.OTPMLimit = otpm
-	k.LimitReset = NormalizeResetWindow(k.LimitReset)
-	k.AllowedModels = decodeModelList(allowed)
-	k.ExpiresAt = expiresAt
-	k.LastUsedAt = lastUsedAt
-	return &k, nil
-}
-
-// encodeModelList serializes a model allow-list for storage. Empty → "".
-func encodeModelList(models []string) string {
-	if len(models) == 0 {
-		return ""
-	}
-	b, err := json.Marshal(models)
-	if err != nil {
-		return ""
-	}
-	return string(b)
-}
-
-// decodeModelList parses a stored model allow-list. "" / invalid → nil.
-func decodeModelList(s string) []string {
-	if s == "" || s == "[]" {
-		return nil
-	}
-	var out []string
-	if err := json.Unmarshal([]byte(s), &out); err != nil {
-		return nil
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-// insertAPIKey writes a fully-formed key record. Shared by CreateAPIKey/SeedKey.
-func (s *PostgresStore) insertAPIKey(ctx context.Context, rec *APIKey, onConflictDoNothing bool) error {
-	q := `INSERT INTO api_keys
-		(id, key_hash, raw_prefix, owner_account_id, name, active,
-		 limit_micro_usd, limit_reset, rpm_limit, itpm_limit, otpm_limit,
-		 allowed_models, expires_at, created_at)
-		VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14)`
-	if onConflictDoNothing {
-		q += ` ON CONFLICT (key_hash) DO NOTHING`
-	}
-	_, err := s.pool.Exec(ctx, q,
-		rec.ID, rec.KeyHash, rec.Label, rec.OwnerAccountID, rec.Name, !rec.Disabled,
-		rec.LimitMicroUSD, NormalizeResetWindow(rec.LimitReset), rec.RPMLimit, rec.ITPMLimit, rec.OTPMLimit,
-		encodeModelList(rec.AllowedModels), rec.ExpiresAt, rec.CreatedAt,
-	)
-	return err
-}
-
-// CreateKey generates a cryptographically random API key, hashes it, stores
-// the hash, and returns the raw key (the only time it's available in plaintext).
-func (s *PostgresStore) CreateKey() (string, error) {
-	raw, _, err := s.CreateAPIKey("", APIKeyCreate{})
-	return raw, err
-}
-
-// CreateKeyForAccount generates a new API key linked to a specific account.
-func (s *PostgresStore) CreateKeyForAccount(accountID string) (string, error) {
-	raw, _, err := s.CreateAPIKey(accountID, APIKeyCreate{})
-	return raw, err
-}
-
-// CreateAPIKey mints a new API key with optional per-key limits.
-func (s *PostgresStore) CreateAPIKey(accountID string, opts APIKeyCreate) (string, *APIKey, error) {
-	raw, err := GenerateRawKey()
-	if err != nil {
-		return "", nil, fmt.Errorf("store: generate key: %w", err)
-	}
-	id, err := GenerateKeyID()
-	if err != nil {
-		return "", nil, fmt.Errorf("store: generate key id: %w", err)
-	}
-	rec := &APIKey{
-		ID:             id,
-		OwnerAccountID: accountID,
-		Name:           opts.Name,
-		Label:          KeyLabel(raw),
-		KeyHash:        HashKey(raw),
-		LimitMicroUSD:  opts.LimitMicroUSD,
-		LimitReset:     NormalizeResetWindow(opts.LimitReset),
-		RPMLimit:       opts.RPMLimit,
-		ITPMLimit:      opts.ITPMLimit,
-		OTPMLimit:      opts.OTPMLimit,
-		AllowedModels:  opts.AllowedModels,
-		ExpiresAt:      opts.ExpiresAt,
-		CreatedAt:      time.Now().UTC(),
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	if err := s.insertAPIKey(ctx, rec, false); err != nil {
-		return "", nil, fmt.Errorf("store: insert key: %w", err)
-	}
-	return raw, rec, nil
-}
-
-// SeedKey inserts a specific raw key into the database. This is used for
-// bootstrapping the admin key. If the key already exists, it is a no-op.
-func (s *PostgresStore) SeedKey(rawKey string) error {
-	id, err := GenerateKeyID()
-	if err != nil {
-		return fmt.Errorf("store: generate key id: %w", err)
-	}
-	rec := &APIKey{
-		ID:         id,
-		Name:       "admin",
-		Label:      KeyLabel(rawKey),
-		KeyHash:    HashKey(rawKey),
-		LimitReset: KeyResetNone,
-		CreatedAt:  time.Now().UTC(),
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	if err := s.insertAPIKey(ctx, rec, true); err != nil {
-		return fmt.Errorf("store: seed key: %w", err)
-	}
-	return nil
-}
-
-// GetKeyAccount returns the account ID that owns this key, or "" if unlinked.
-func (s *PostgresStore) GetKeyAccount(key string) string {
-	h := HashKey(key)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var accountID string
-	err := s.pool.QueryRow(ctx,
-		`SELECT owner_account_id FROM api_keys WHERE key_hash = $1 AND active = TRUE`, h,
-	).Scan(&accountID)
-	if err != nil {
-		return ""
-	}
-	return accountID
-}
-
-// ValidateKey returns true if the given key exists, is active, and is not
-// expired. Expiry is enforced here (not just in AuthenticateKey) so callers
-// like telemetry attribution don't treat an expired key as a live account.
-func (s *PostgresStore) ValidateKey(key string) bool {
-	h := HashKey(key)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var active bool
-	var expiresAt *time.Time
-	err := s.pool.QueryRow(ctx,
-		`SELECT active, expires_at FROM api_keys WHERE key_hash = $1`,
-		h,
-	).Scan(&active, &expiresAt)
-	if err != nil {
-		return false
-	}
-	if expiresAt != nil && time.Now().After(*expiresAt) {
-		return false
-	}
-	return active
-}
-
-// ValidateKeyFull returns the active status and owner account ID for an
-// API key in a single query. Returns an error if the key does not exist.
-func (s *PostgresStore) ValidateKeyFull(key string) (bool, string, error) {
-	h := HashKey(key)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var active bool
-	var ownerAccountID string
-	err := s.pool.QueryRow(ctx,
-		`SELECT active, owner_account_id FROM api_keys WHERE key_hash = $1`, h,
-	).Scan(&active, &ownerAccountID)
-	if err != nil {
-		return false, "", err
-	}
-	return active, ownerAccountID, nil
-}
-
-// AuthenticateKey resolves a raw key to its active record for request auth.
-func (s *PostgresStore) AuthenticateKey(rawKey string) (*APIKey, error) {
-	h := HashKey(rawKey)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+apiKeyColumns+` FROM api_keys WHERE key_hash = $1`, h)
-	k, err := scanAPIKeyRow(row)
-	if err != nil {
-		return nil, err
-	}
-	if k.Disabled {
-		return nil, fmt.Errorf("key disabled")
-	}
-	if k.ExpiresAt != nil && time.Now().After(*k.ExpiresAt) {
-		return nil, fmt.Errorf("key expired")
-	}
-	return k, nil
-}
-
-// ListAPIKeys returns all keys owned by an account, newest first.
-func (s *PostgresStore) ListAPIKeys(accountID string) ([]APIKey, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT `+apiKeyColumns+` FROM api_keys WHERE owner_account_id = $1 AND id <> '' ORDER BY created_at DESC`, accountID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	out := make([]APIKey, 0)
-	for rows.Next() {
-		k, err := scanAPIKeyRow(rows)
-		if err != nil {
-			return nil, err
-		}
-		out = append(out, *k)
-	}
-	return out, rows.Err()
-}
-
-// GetAPIKeyByID returns a single key by ID, scoped to the owner.
-func (s *PostgresStore) GetAPIKeyByID(accountID, id string) (*APIKey, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+apiKeyColumns+` FROM api_keys WHERE id = $1 AND owner_account_id = $2`, id, accountID)
-	return scanAPIKeyRow(row)
-}
-
-// UpdateAPIKey overwrites mutable fields of a key, scoped to the owner.
-func (s *PostgresStore) UpdateAPIKey(accountID, id string, mutable APIKey) (*APIKey, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE api_keys SET
-			name = $1, active = $2, limit_micro_usd = $3, limit_reset = $4,
-			rpm_limit = $5, itpm_limit = $6, otpm_limit = $7,
-			allowed_models = $8, expires_at = $9
-		 WHERE id = $10 AND owner_account_id = $11`,
-		mutable.Name, !mutable.Disabled, mutable.LimitMicroUSD, NormalizeResetWindow(mutable.LimitReset),
-		mutable.RPMLimit, mutable.ITPMLimit, mutable.OTPMLimit,
-		encodeModelList(mutable.AllowedModels), mutable.ExpiresAt,
-		id, accountID,
-	)
-	if err != nil {
-		return nil, err
-	}
-	if tag.RowsAffected() == 0 {
-		return nil, fmt.Errorf("key not found")
-	}
-	return s.GetAPIKeyByID(accountID, id)
-}
-
-// RevokeAPIKeyByID permanently deletes a key by ID, scoped to the owner.
-func (s *PostgresStore) RevokeAPIKeyByID(accountID, id string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`DELETE FROM api_keys WHERE id = $1 AND owner_account_id = $2`, id, accountID)
-	if err != nil {
-		return err
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("key not found")
-	}
-	return nil
-}
-
-// RotateAPIKey atomically replaces a key within a transaction (see Store
-// interface). The old key is deleted and the new key inserted in the same tx;
-// a concurrent rotate of the same id finds the row gone and returns not-found.
-func (s *PostgresStore) RotateAPIKey(accountID, id string) (string, *APIKey, error) {
-	raw, err := GenerateRawKey()
-	if err != nil {
-		return "", nil, fmt.Errorf("store: generate key: %w", err)
-	}
-	newID, err := GenerateKeyID()
-	if err != nil {
-		return "", nil, fmt.Errorf("store: generate key id: %w", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return "", nil, fmt.Errorf("store: begin rotate tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	old, err := scanAPIKeyRow(tx.QueryRow(ctx,
-		`SELECT `+apiKeyColumns+` FROM api_keys WHERE id = $1 AND owner_account_id = $2 FOR UPDATE`, id, accountID))
-	if err != nil {
-		return "", nil, fmt.Errorf("key not found")
-	}
-
-	rec := &APIKey{
-		ID:             newID,
-		OwnerAccountID: accountID,
-		Name:           old.Name,
-		Label:          KeyLabel(raw),
-		KeyHash:        HashKey(raw),
-		Disabled:       old.Disabled,
-		LimitMicroUSD:  old.LimitMicroUSD,
-		LimitReset:     NormalizeResetWindow(old.LimitReset),
-		RPMLimit:       old.RPMLimit,
-		ITPMLimit:      old.ITPMLimit,
-		OTPMLimit:      old.OTPMLimit,
-		AllowedModels:  old.AllowedModels,
-		ExpiresAt:      old.ExpiresAt,
-		CreatedAt:      time.Now().UTC(),
-	}
-	if _, err := tx.Exec(ctx,
-		`INSERT INTO api_keys
-			(id, key_hash, raw_prefix, owner_account_id, name, active,
-			 limit_micro_usd, limit_reset, rpm_limit, itpm_limit, otpm_limit,
-			 allowed_models, expires_at, created_at)
-		 VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14)`,
-		rec.ID, rec.KeyHash, rec.Label, rec.OwnerAccountID, rec.Name, !rec.Disabled,
-		rec.LimitMicroUSD, rec.LimitReset, rec.RPMLimit, rec.ITPMLimit, rec.OTPMLimit,
-		encodeModelList(rec.AllowedModels), rec.ExpiresAt, rec.CreatedAt,
-	); err != nil {
-		return "", nil, fmt.Errorf("store: insert rotated key: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM api_keys WHERE id = $1 AND owner_account_id = $2`, id, accountID); err != nil {
-		return "", nil, fmt.Errorf("store: delete old key: %w", err)
-	}
-	if err := tx.Commit(ctx); err != nil {
-		return "", nil, fmt.Errorf("store: commit rotate: %w", err)
-	}
-	return raw, rec, nil
-}
-
-// TouchAPIKey records that a key was used at the given time.
-func (s *PostgresStore) TouchAPIKey(id string, at time.Time) {
-	if id == "" {
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	_, _ = s.pool.Exec(ctx, `UPDATE api_keys SET last_used_at = $1 WHERE id = $2`, at.UTC(), id)
-}
-
-// KeySpendSince returns total micro-USD charged to a key since `since` (UTC).
-func (s *PostgresStore) KeySpendSince(keyID string, since time.Time) int64 {
-	if keyID == "" {
-		return 0
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var total int64
-	err := s.pool.QueryRow(ctx,
-		`SELECT COALESCE(SUM(cost_micro_usd), 0) FROM usage
-		 WHERE key_id = $1 AND ($2::timestamptz IS NULL OR created_at >= $2)`,
-		keyID, nullSince(since),
-	).Scan(&total)
-	if err != nil {
-		return 0
-	}
-	return total
-}
-
-// RevokeKey deactivates a key. Returns true if the key existed and was active.
-func (s *PostgresStore) RevokeKey(key string) bool {
-	h := HashKey(key)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE api_keys SET active = FALSE WHERE key_hash = $1 AND active = TRUE`,
-		h,
-	)
-	if err != nil {
-		return false
-	}
-	return tag.RowsAffected() > 0
-}
-
-// RecordUsage inserts a usage record into PostgreSQL.
-func (s *PostgresStore) RecordUsage(providerID, consumerKey, model string, promptTokens, completionTokens int) {
-	h := HashKey(consumerKey)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, _ = s.pool.Exec(ctx,
-		`WITH ins AS (
-			INSERT INTO usage (provider_id, consumer_key_hash, model, prompt_tokens, completion_tokens)
-			VALUES ($1, $2, $3, $4, $5)
-		)
-		UPDATE usage_totals SET
-			total_requests = total_requests + 1,
-			total_prompt_tokens = total_prompt_tokens + $4,
-			total_completion_tokens = total_completion_tokens + $5
-		WHERE id = 1`,
-		providerID, h, model, promptTokens, completionTokens,
-	)
-}
-
-// UsageByConsumer returns usage records for a specific consumer key.
-func (s *PostgresStore) UsageByConsumer(consumerKey string) []UsageRecord {
-	h := HashKey(consumerKey)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT provider_id, consumer_key_hash, model, prompt_tokens, completion_tokens, created_at, request_id, cost_micro_usd
-		 FROM usage WHERE consumer_key_hash = $1 ORDER BY created_at DESC LIMIT 100`, h)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var records []UsageRecord
-	for rows.Next() {
-		var r UsageRecord
-		if err := rows.Scan(&r.ProviderID, &r.ConsumerKey, &r.Model, &r.PromptTokens, &r.CompletionTokens, &r.CreatedAt, &r.RequestID, &r.CostMicroUSD); err != nil {
-			continue
-		}
-		records = append(records, r)
-	}
-	return records
-}
-
-// RecordUsageWithCost inserts a usage record with request ID and cost.
-func (s *PostgresStore) RecordUsageWithCost(providerID, consumerKey, model, requestID string, promptTokens, completionTokens int, costMicroUSD int64) {
-	s.RecordUsageWithCostAndLocation(providerID, consumerKey, model, requestID, promptTokens, completionTokens, costMicroUSD, nil)
-}
-
-// RecordUsageWithCostAndLocation inserts a usage record with request ID, cost,
-// and approximate request-origin location.
-func (s *PostgresStore) RecordUsageWithCostAndLocation(providerID, consumerKey, model, requestID string, promptTokens, completionTokens int, costMicroUSD int64, requestLocation *ProviderLocation) {
-	s.RecordUsageFull(providerID, consumerKey, "", model, requestID, promptTokens, completionTokens, costMicroUSD, requestLocation)
-}
-
-// RecordUsageFull inserts a usage record with full attribution including the
-// originating API key ID for per-key usage and spend tracking.
-func (s *PostgresStore) RecordUsageFull(providerID, consumerKey, keyID, model, requestID string, promptTokens, completionTokens int, costMicroUSD int64, requestLocation *ProviderLocation) {
-	h := HashKey(consumerKey)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, _ = s.pool.Exec(ctx,
-		`WITH ins AS (
-			INSERT INTO usage (provider_id, consumer_key_hash, key_id, model, prompt_tokens, completion_tokens, request_id, cost_micro_usd, request_location)
-			VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
-		)
-		UPDATE usage_totals SET
-			total_requests = total_requests + 1,
-			total_prompt_tokens = total_prompt_tokens + $5,
-			total_completion_tokens = total_completion_tokens + $6
-		WHERE id = 1`,
-		providerID, h, keyID, model, promptTokens, completionTokens, requestID, costMicroUSD, marshalProviderLocation(requestLocation),
-	)
-}
-
-// UsageLocationBuckets aggregates usage by approximate request origin.
-func (s *PostgresStore) UsageLocationBuckets(since time.Time) []UsageLocationBucket {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT
-			COALESCE(request_location->>'city', '') AS city,
-			COALESCE(request_location->>'region', '') AS region,
-			COALESCE(request_location->>'region_code', '') AS region_code,
-			COALESCE(request_location->>'country', '') AS country,
-			COALESCE(request_location->>'country_code', '') AS country_code,
-			COALESCE(AVG(NULLIF(request_location->>'latitude', '')::double precision), 0),
-			COALESCE(AVG(NULLIF(request_location->>'longitude', '')::double precision), 0),
-			COUNT(*),
-			COALESCE(SUM(prompt_tokens), 0),
-			COALESCE(SUM(completion_tokens), 0),
-			COUNT(DISTINCT provider_id)
-		 FROM usage
-		 WHERE request_location IS NOT NULL
-		   AND ($1::timestamptz IS NULL OR created_at >= $1)
-		 GROUP BY city, region, region_code, country, country_code
-		 ORDER BY COUNT(*) DESC`,
-		nullSince(since),
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var buckets []UsageLocationBucket
-	for rows.Next() {
-		var b UsageLocationBucket
-		if err := rows.Scan(
-			&b.City,
-			&b.Region,
-			&b.RegionCode,
-			&b.Country,
-			&b.CountryCode,
-			&b.Latitude,
-			&b.Longitude,
-			&b.Requests,
-			&b.PromptTokens,
-			&b.CompletionTokens,
-			&b.Providers,
-		); err != nil {
-			continue
-		}
-		buckets = append(buckets, b)
-	}
-	return buckets
-}
-
-// UsageFlowBuckets aggregates directional consumer→provider flows by JOINing
-// the usage table with providers in SQL. This replaces loading all rows into
-// Go and doing the aggregation in-process. The query only returns the top 50
-// flows (by request count) so the result set is bounded.
-func (s *PostgresStore) UsageFlowBuckets(since time.Time, _ map[string]*ProviderLocation) []UsageFlowBucket {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT
-			COALESCE(u.request_location->>'city', '')         AS c_city,
-			COALESCE(u.request_location->>'region', '')       AS c_region,
-			COALESCE(u.request_location->>'region_code', '')  AS c_region_code,
-			COALESCE(u.request_location->>'country', '')      AS c_country,
-			COALESCE(u.request_location->>'country_code', '') AS c_country_code,
-			COALESCE(AVG(NULLIF(u.request_location->>'latitude',  '')::double precision), 0) AS c_lat,
-			COALESCE(AVG(NULLIF(u.request_location->>'longitude', '')::double precision), 0) AS c_lng,
-			COALESCE(p.location->>'city', '')         AS p_city,
-			COALESCE(p.location->>'region', '')       AS p_region,
-			COALESCE(p.location->>'region_code', '')  AS p_region_code,
-			COALESCE(p.location->>'country', '')      AS p_country,
-			COALESCE(p.location->>'country_code', '') AS p_country_code,
-			COALESCE(AVG(NULLIF(p.location->>'latitude',  '')::double precision), 0) AS p_lat,
-			COALESCE(AVG(NULLIF(p.location->>'longitude', '')::double precision), 0) AS p_lng,
-			COUNT(*)                              AS requests,
-			COALESCE(SUM(u.prompt_tokens), 0)     AS prompt_tokens,
-			COALESCE(SUM(u.completion_tokens), 0) AS completion_tokens
-		 FROM usage u
-		 JOIN providers p ON p.id = u.provider_id
-		 WHERE u.request_location IS NOT NULL
-		   AND p.location IS NOT NULL
-		   AND ($1::timestamptz IS NULL OR u.created_at >= $1)
-		 GROUP BY c_city, c_region, c_region_code, c_country, c_country_code,
-		          p_city, p_region, p_region_code, p_country, p_country_code
-		 ORDER BY requests DESC
-		 LIMIT 50`,
-		nullSince(since),
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var buckets []UsageFlowBucket
-	for rows.Next() {
-		var b UsageFlowBucket
-		if err := rows.Scan(
-			&b.ConsumerCity, &b.ConsumerRegion, &b.ConsumerRegionCode,
-			&b.ConsumerCountry, &b.ConsumerCountryCode,
-			&b.ConsumerLatitude, &b.ConsumerLongitude,
-			&b.ProviderCity, &b.ProviderRegion, &b.ProviderRegionCode,
-			&b.ProviderCountry, &b.ProviderCountryCode,
-			&b.ProviderLatitude, &b.ProviderLongitude,
-			&b.Requests, &b.PromptTokens, &b.CompletionTokens,
-		); err != nil {
-			continue
-		}
-		buckets = append(buckets, b)
-	}
-	return buckets
-}
-
-func nullSince(since time.Time) any {
-	if since.IsZero() {
-		return nil
-	}
-	return since
-}
-
-// RecordPayment inserts a payment record into PostgreSQL.
-func (s *PostgresStore) RecordPayment(txHash, consumerAddr, providerAddr, amountUSD, model string, promptTokens, completionTokens int, memo string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO payments (tx_hash, consumer_address, provider_address, amount_usd, model, prompt_tokens, completion_tokens, memo)
-		 VALUES ($1, $2, $3, $4, $5, $6, $7, $8)`,
-		txHash, consumerAddr, providerAddr, amountUSD, model, promptTokens, completionTokens, memo,
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert payment: %w", err)
-	}
-	return nil
-}
-
-// UsageCountSince returns the number of usage records created at or after the
-// given time. Uses idx_usage_created for an index-only count.
-func (s *PostgresStore) UsageCountSince(since time.Time) int64 {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var count int64
-	_ = s.pool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM usage
-		 WHERE ($1::timestamptz IS NULL OR created_at >= $1)`,
-		nullSince(since),
-	).Scan(&count)
-	return count
-}
-
-// UsageTotals returns aggregated lifetime totals from the materialized
-// usage_totals counter row. This is a single PK lookup — O(1) regardless
-// of how many rows exist in the usage table.
-func (s *PostgresStore) UsageTotals() UsageTotals {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var t UsageTotals
-	_ = s.pool.QueryRow(ctx,
-		`SELECT total_requests, total_prompt_tokens, total_completion_tokens
-		 FROM usage_totals WHERE id = 1`,
-	).Scan(&t.Requests, &t.PromptTokens, &t.CompletionTokens)
-	return t
-}
-
-// UsageTimeSeries returns per-minute usage buckets at or after `since`.
-func (s *PostgresStore) UsageTimeSeries(since time.Time) []UsageBucket {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT date_trunc('minute', created_at) AS minute,
-		        COUNT(*),
-		        COALESCE(SUM(prompt_tokens), 0),
-		        COALESCE(SUM(completion_tokens), 0)
-		 FROM usage
-		 WHERE created_at >= $1
-		 GROUP BY minute
-		 ORDER BY minute ASC`,
-		since,
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var buckets []UsageBucket
-	for rows.Next() {
-		var b UsageBucket
-		if err := rows.Scan(&b.Minute, &b.Requests, &b.PromptTokens, &b.CompletionTokens); err != nil {
-			continue
-		}
-		buckets = append(buckets, b)
-	}
-	return buckets
-}
-
-// Leaderboard returns the top N accounts ranked by the given metric over the
-// given time window. Zero `since` means all-time. The ranking is computed in
-// SQL via aggregation on provider_earnings — no per-row wire transfer.
-func (s *PostgresStore) Leaderboard(metric LeaderboardMetric, since time.Time, limit int) []LeaderboardRow {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	if limit <= 0 || limit > 200 {
-		limit = 50
-	}
-
-	orderBy := "earnings_micro_usd DESC"
-	switch metric {
-	case LeaderboardTokens:
-		orderBy = "tokens DESC"
-	case LeaderboardJobs:
-		orderBy = "jobs DESC"
-	}
-
-	// account_id != '' filters out unassigned earnings (e.g. legacy wallet-only).
-	q := `SELECT account_id,
-	             COALESCE(SUM(amount_micro_usd), 0)               AS earnings_micro_usd,
-	             COALESCE(SUM(prompt_tokens + completion_tokens), 0) AS tokens,
-	             COUNT(*)                                          AS jobs
-	      FROM provider_earnings
-	      WHERE account_id != ''`
-	args := []any{}
-	if !since.IsZero() {
-		q += ` AND created_at >= $1`
-		args = append(args, since)
-	}
-	q += `
-	      GROUP BY account_id
-	      ORDER BY ` + orderBy + `
-	      LIMIT $` + strconv.Itoa(len(args)+1)
-	args = append(args, limit)
-
-	rows, err := s.pool.Query(ctx, q, args...)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	out := make([]LeaderboardRow, 0, limit)
-	for rows.Next() {
-		var r LeaderboardRow
-		if err := rows.Scan(&r.AccountID, &r.EarningsMicroUSD, &r.Tokens, &r.Jobs); err != nil {
-			continue
-		}
-		out = append(out, r)
-	}
-	return out
-}
-
-// NetworkTotals returns aggregated metrics across all earnings for the given
-// time window. Zero `since` means all-time.
-func (s *PostgresStore) NetworkTotals(since time.Time) NetworkTotalsRow {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	q := `SELECT COALESCE(SUM(amount_micro_usd), 0),
-	             COALESCE(SUM(prompt_tokens + completion_tokens), 0),
-	             COUNT(*),
-	             COUNT(DISTINCT account_id) FILTER (WHERE account_id != '')
-	      FROM provider_earnings`
-	args := []any{}
-	if !since.IsZero() {
-		q += ` WHERE created_at >= $1`
-		args = append(args, since)
-	}
-
-	var t NetworkTotalsRow
-	_ = s.pool.QueryRow(ctx, q, args...).
-		Scan(&t.EarningsMicroUSD, &t.Tokens, &t.Jobs, &t.ActiveAccounts)
-	return t
-}
-
-// UsageRecords returns usage records from the database, ordered by creation time.
-// Limited to the most recent 10000 rows as a safety guard against unbounded reads.
-func (s *PostgresStore) UsageRecords() []UsageRecord {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT provider_id, consumer_key_hash, model, prompt_tokens, completion_tokens, created_at, request_id, cost_micro_usd, request_location
-		 FROM usage ORDER BY created_at DESC LIMIT 10000`,
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var records []UsageRecord
-	for rows.Next() {
-		var r UsageRecord
-		var locationRaw []byte
-		if err := rows.Scan(
-			&r.ProviderID,
-			&r.ConsumerKey,
-			&r.Model,
-			&r.PromptTokens,
-			&r.CompletionTokens,
-			&r.Timestamp,
-			&r.RequestID,
-			&r.CostMicroUSD,
-			&locationRaw,
-		); err != nil {
-			continue
-		}
-		r.CreatedAt = r.Timestamp
-		r.RequestLocation = unmarshalProviderLocation(locationRaw)
-		records = append(records, r)
-	}
-	if records == nil {
-		records = make([]UsageRecord, 0)
-	}
-	return records
-}
-
-// UsageRecordsSince returns usage records created at or after the given time.
-func (s *PostgresStore) UsageRecordsSince(since time.Time) []UsageRecord {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT provider_id, consumer_key_hash, model, prompt_tokens, completion_tokens, created_at, request_id, cost_micro_usd, request_location
-		 FROM usage
-		 WHERE ($1::timestamptz IS NULL OR created_at >= $1)
-		 ORDER BY created_at ASC`,
-		nullSince(since),
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var records []UsageRecord
-	for rows.Next() {
-		var r UsageRecord
-		var locationRaw []byte
-		if err := rows.Scan(
-			&r.ProviderID,
-			&r.ConsumerKey,
-			&r.Model,
-			&r.PromptTokens,
-			&r.CompletionTokens,
-			&r.Timestamp,
-			&r.RequestID,
-			&r.CostMicroUSD,
-			&locationRaw,
-		); err != nil {
-			continue
-		}
-		r.CreatedAt = r.Timestamp
-		r.RequestLocation = unmarshalProviderLocation(locationRaw)
-		records = append(records, r)
-	}
-	if records == nil {
-		return []UsageRecord{}
-	}
-	return records
-}
-
-// GetBalance returns the current balance in micro-USD for an account.
-func (s *PostgresStore) GetBalance(accountID string) int64 {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var balance int64
-	err := s.pool.QueryRow(ctx,
-		`SELECT balance_micro_usd FROM balances WHERE account_id = $1`, accountID,
-	).Scan(&balance)
-	if err != nil {
-		return 0
-	}
-	return balance
-}
-
-func nullableCreatedAt(ts time.Time) any {
-	if ts.IsZero() {
-		return nil
-	}
-	return ts
-}
-
-func creditTx(ctx context.Context, tx pgx.Tx, accountID string, amountMicroUSD int64, entryType LedgerEntryType, reference string, createdAt time.Time) error {
-	_, err := tx.Exec(ctx,
-		`INSERT INTO balances (account_id, balance_micro_usd, updated_at)
-		 VALUES ($1, $2, NOW())
-		 ON CONFLICT (account_id) DO UPDATE SET
-		   balance_micro_usd = balances.balance_micro_usd + $2,
-		   updated_at = NOW()`,
-		accountID, amountMicroUSD,
-	)
-	if err != nil {
-		return fmt.Errorf("store: credit balance: %w", err)
-	}
-
-	var balanceAfter int64
-	err = tx.QueryRow(ctx,
-		`SELECT balance_micro_usd FROM balances WHERE account_id = $1`, accountID,
-	).Scan(&balanceAfter)
-	if err != nil {
-		return fmt.Errorf("store: read balance: %w", err)
-	}
-
-	_, err = tx.Exec(ctx,
-		`INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference, created_at)
-		 VALUES ($1, $2, $3, $4, $5, COALESCE($6, NOW()))`,
-		accountID, string(entryType), amountMicroUSD, balanceAfter, reference, nullableCreatedAt(createdAt),
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert ledger entry: %w", err)
-	}
-
-	return nil
-}
-
-func creditWithdrawableTx(ctx context.Context, tx pgx.Tx, accountID string, amountMicroUSD int64, entryType LedgerEntryType, reference string, createdAt time.Time) error {
-	_, err := tx.Exec(ctx,
-		`INSERT INTO balances (account_id, balance_micro_usd, withdrawable_micro_usd, updated_at)
-		 VALUES ($1, $2, $2, NOW())
-		 ON CONFLICT (account_id) DO UPDATE SET
-		   balance_micro_usd = balances.balance_micro_usd + $2,
-		   withdrawable_micro_usd = balances.withdrawable_micro_usd + $2,
-		   updated_at = NOW()`,
-		accountID, amountMicroUSD,
-	)
-	if err != nil {
-		return fmt.Errorf("store: credit withdrawable balance: %w", err)
-	}
-
-	var balanceAfter int64
-	err = tx.QueryRow(ctx,
-		`SELECT balance_micro_usd FROM balances WHERE account_id = $1`, accountID,
-	).Scan(&balanceAfter)
-	if err != nil {
-		return fmt.Errorf("store: read balance: %w", err)
-	}
-
-	_, err = tx.Exec(ctx,
-		`INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference, created_at)
-		 VALUES ($1, $2, $3, $4, $5, COALESCE($6, NOW()))`,
-		accountID, string(entryType), amountMicroUSD, balanceAfter, reference, nullableCreatedAt(createdAt),
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert ledger entry: %w", err)
-	}
-
-	return nil
-}
-
-// Credit adds micro-USD to an account and records a ledger entry (atomic).
-func (s *PostgresStore) Credit(accountID string, amountMicroUSD int64, entryType LedgerEntryType, reference string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("store: begin tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	if err := creditTx(ctx, tx, accountID, amountMicroUSD, entryType, reference, time.Time{}); err != nil {
-		return err
-	}
-
-	return tx.Commit(ctx)
-}
-
-// GetWithdrawableBalance returns the withdrawable balance in micro-USD.
-func (s *PostgresStore) GetWithdrawableBalance(accountID string) int64 {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var balance int64
-	err := s.pool.QueryRow(ctx,
-		`SELECT withdrawable_micro_usd FROM balances WHERE account_id = $1`, accountID,
-	).Scan(&balance)
-	if err != nil {
-		return 0
-	}
-	return balance
-}
-
-// GetBalanceWithWithdrawable returns both balances in a single query.
-func (s *PostgresStore) GetBalanceWithWithdrawable(accountID string) (int64, int64) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var balance, withdrawable int64
-	err := s.pool.QueryRow(ctx,
-		`SELECT balance_micro_usd, withdrawable_micro_usd FROM balances WHERE account_id = $1`, accountID,
-	).Scan(&balance, &withdrawable)
-	if err != nil {
-		return 0, 0
-	}
-	return balance, withdrawable
-}
-
-// CreditWithdrawable adds micro-USD to both the total balance and the
-// withdrawable balance, and records a ledger entry.
-func (s *PostgresStore) CreditWithdrawable(accountID string, amountMicroUSD int64, entryType LedgerEntryType, reference string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("store: begin tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	if err := creditWithdrawableTx(ctx, tx, accountID, amountMicroUSD, entryType, reference, time.Time{}); err != nil {
-		return err
-	}
-
-	return tx.Commit(ctx)
-}
-
-// Debit subtracts micro-USD from an account. Returns error if insufficient funds.
-func (s *PostgresStore) Debit(accountID string, amountMicroUSD int64, entryType LedgerEntryType, reference string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	// Single-statement CTE: debit balance, cap withdrawable, insert ledger
-	// entry -- all in one round trip. The old implementation used 5 sequential
-	// round trips (BEGIN + 2 UPDATEs + INSERT + COMMIT) which paid full
-	// network latency to Postgres on each hop (~200ms × 5 = 1s+).
-	var balanceAfter int64
-	err := s.pool.QueryRow(ctx, `
-		WITH debit AS (
-			UPDATE balances
-			SET balance_micro_usd = balance_micro_usd - $2,
-			    withdrawable_micro_usd = LEAST(withdrawable_micro_usd, balance_micro_usd - $2),
-			    updated_at = NOW()
-			WHERE account_id = $1 AND balance_micro_usd >= $2
-			RETURNING balance_micro_usd
-		), ledger AS (
-			INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference)
-			SELECT $1, $3, -$2, balance_micro_usd, $4
-			FROM debit
-		)
-		SELECT balance_micro_usd FROM debit`,
-		accountID, amountMicroUSD, string(entryType), reference,
-	).Scan(&balanceAfter)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return ErrInsufficientBalance
-		}
-		return fmt.Errorf("debit: %w", err)
-	}
-	return nil
-}
-
-// MigrateAccountBalance moves the full balance (and withdrawable subset) from
-// one account ID to another in a single transaction. No-op (false) when the
-// source has no balance row or a zero balance.
-func (s *PostgresStore) MigrateAccountBalance(from, to string) (bool, error) {
-	if from == "" || to == "" || from == to {
-		return false, nil
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return false, fmt.Errorf("store: begin migrate tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	var bal, wdr int64
-	err = tx.QueryRow(ctx,
-		`SELECT balance_micro_usd, withdrawable_micro_usd FROM balances WHERE account_id = $1 FOR UPDATE`, from,
-	).Scan(&bal, &wdr)
-	if errors.Is(err, pgx.ErrNoRows) {
-		return false, nil
-	}
-	if err != nil {
-		return false, fmt.Errorf("store: read source balance: %w", err)
-	}
-	if bal == 0 && wdr == 0 {
-		return false, nil
-	}
-
-	// Zero the source and record the outgoing leg.
-	if _, err := tx.Exec(ctx,
-		`UPDATE balances SET balance_micro_usd = 0, withdrawable_micro_usd = 0, updated_at = NOW() WHERE account_id = $1`, from,
-	); err != nil {
-		return false, fmt.Errorf("store: zero source balance: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference)
-		 VALUES ($1, $2, $3, 0, 'migrate:out')`,
-		from, string(LedgerMigration), -bal,
-	); err != nil {
-		return false, fmt.Errorf("store: source migration ledger entry: %w", err)
-	}
-
-	// Credit the destination and record the incoming leg.
-	var destBalance int64
-	if err := tx.QueryRow(ctx,
-		`INSERT INTO balances (account_id, balance_micro_usd, withdrawable_micro_usd, updated_at)
-		 VALUES ($1, $2, $3, NOW())
-		 ON CONFLICT (account_id) DO UPDATE SET
-		   balance_micro_usd = balances.balance_micro_usd + $2,
-		   withdrawable_micro_usd = balances.withdrawable_micro_usd + $3,
-		   updated_at = NOW()
-		 RETURNING balance_micro_usd`,
-		to, bal, wdr,
-	).Scan(&destBalance); err != nil {
-		return false, fmt.Errorf("store: credit destination balance: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference)
-		 VALUES ($1, $2, $3, $4, 'migrate:in')`,
-		to, string(LedgerMigration), bal, destBalance,
-	); err != nil {
-		return false, fmt.Errorf("store: destination migration ledger entry: %w", err)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return false, fmt.Errorf("store: commit migrate: %w", err)
-	}
-	return true, nil
-}
-
-// DebitWithdrawable subtracts micro-USD from both the total balance and the
-// withdrawable balance atomically. Returns error if the withdrawable balance
-// is insufficient. This ensures withdrawal debits are symmetric with
-// CreditWithdrawable refunds — both touch the same columns.
-func (s *PostgresStore) DebitWithdrawable(accountID string, amountMicroUSD int64, entryType LedgerEntryType, reference string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("store: begin tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	var balanceAfter int64
-	err = tx.QueryRow(ctx,
-		`UPDATE balances
-		 SET balance_micro_usd = balance_micro_usd - $2,
-		     withdrawable_micro_usd = withdrawable_micro_usd - $2,
-		     updated_at = NOW()
-		 WHERE account_id = $1
-		   AND balance_micro_usd >= $2
-		   AND withdrawable_micro_usd >= $2
-		 RETURNING balance_micro_usd`,
-		accountID, amountMicroUSD,
-	).Scan(&balanceAfter)
-	if err != nil {
-		return errors.New("insufficient withdrawable balance or account not found")
-	}
-
-	_, err = tx.Exec(ctx,
-		`INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference)
-		 VALUES ($1, $2, $3, $4, $5)`,
-		accountID, string(entryType), -amountMicroUSD, balanceAfter, reference,
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert ledger entry: %w", err)
-	}
-
-	return tx.Commit(ctx)
-}
-
-// LedgerHistory returns ledger entries for an account, newest first.
-func (s *PostgresStore) LedgerHistory(accountID string) []LedgerEntry {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	// Cap at 500 most-recent entries. Older history isn't shown on any
-	// dashboard and was responsible for sending tens of thousands of rows
-	// per request to high-volume accounts.
-	rows, err := s.pool.Query(ctx,
-		`SELECT id, account_id, entry_type, amount_micro_usd, balance_after, reference, created_at
-		 FROM ledger_entries WHERE account_id = $1 ORDER BY created_at DESC LIMIT 500`,
-		accountID,
-	)
-	if err != nil {
-		return []LedgerEntry{}
-	}
-	defer rows.Close()
-
-	var entries []LedgerEntry
-	for rows.Next() {
-		var e LedgerEntry
-		var entryType string
-		if err := rows.Scan(&e.ID, &e.AccountID, &entryType, &e.AmountMicroUSD, &e.BalanceAfter, &e.Reference, &e.CreatedAt); err != nil {
-			continue
-		}
-		e.Type = LedgerEntryType(entryType)
-		entries = append(entries, e)
-	}
-	if entries == nil {
-		return []LedgerEntry{}
-	}
-	return entries
-}
-
-// KeyCount returns the number of active API keys.
-func (s *PostgresStore) KeyCount() int {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var count int
-	err := s.pool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM api_keys WHERE active = TRUE`,
-	).Scan(&count)
-	if err != nil {
-		return 0
-	}
-	return count
-}
-
-// --- Referral System ---
-
-// CreateReferrer registers an account as a referrer with the given code.
-func (s *PostgresStore) CreateReferrer(accountID, code string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO referrers (account_id, code) VALUES ($1, $2)`,
-		accountID, code,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create referrer: %w", err)
-	}
-	return nil
-}
-
-// GetReferrerByCode returns the referrer for a given referral code.
-func (s *PostgresStore) GetReferrerByCode(code string) (*Referrer, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var ref Referrer
-	err := s.pool.QueryRow(ctx,
-		`SELECT account_id, code, created_at FROM referrers WHERE code = $1`, code,
-	).Scan(&ref.AccountID, &ref.Code, &ref.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: referrer not found: %w", err)
-	}
-	return &ref, nil
-}
-
-// GetReferrerByAccount returns the referrer record for an account.
-func (s *PostgresStore) GetReferrerByAccount(accountID string) (*Referrer, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var ref Referrer
-	err := s.pool.QueryRow(ctx,
-		`SELECT account_id, code, created_at FROM referrers WHERE account_id = $1`, accountID,
-	).Scan(&ref.AccountID, &ref.Code, &ref.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: referrer not found: %w", err)
-	}
-	return &ref, nil
-}
-
-// RecordReferral records that referredAccountID was referred by referrerCode.
-func (s *PostgresStore) RecordReferral(referrerCode, referredAccountID string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO referrals (referred_account, referrer_code) VALUES ($1, $2)`,
-		referredAccountID, referrerCode,
-	)
-	if err != nil {
-		return fmt.Errorf("store: record referral: %w", err)
-	}
-	return nil
-}
-
-// GetReferrerForAccount returns the referrer code that referred this account.
-func (s *PostgresStore) GetReferrerForAccount(accountID string) (string, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var code string
-	err := s.pool.QueryRow(ctx,
-		`SELECT referrer_code FROM referrals WHERE referred_account = $1`, accountID,
-	).Scan(&code)
-	if err != nil {
-		return "", nil // no referrer is not an error
-	}
-	return code, nil
-}
-
-// GetReferralStats returns referral statistics for a code.
-func (s *PostgresStore) GetReferralStats(code string) (*ReferralStats, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	// Verify code exists
-	var accountID string
-	err := s.pool.QueryRow(ctx,
-		`SELECT account_id FROM referrers WHERE code = $1`, code,
-	).Scan(&accountID)
-	if err != nil {
-		return nil, fmt.Errorf("store: referral code not found: %w", err)
-	}
-
-	// Count referred accounts
-	var totalReferred int
-	_ = s.pool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM referrals WHERE referrer_code = $1`, code,
-	).Scan(&totalReferred)
-
-	// Sum referral rewards from ledger
-	var totalRewards int64
-	_ = s.pool.QueryRow(ctx,
-		`SELECT COALESCE(SUM(amount_micro_usd), 0) FROM ledger_entries
-		 WHERE account_id = $1 AND entry_type = $2`,
-		accountID, string(LedgerReferralReward),
-	).Scan(&totalRewards)
-
-	return &ReferralStats{
-		Code:                 code,
-		TotalReferred:        totalReferred,
-		TotalRewardsMicroUSD: totalRewards,
-	}, nil
-}
-
-// --- Billing Sessions ---
-
-// CreateBillingSession stores a new billing session.
-func (s *PostgresStore) CreateBillingSession(session *BillingSession) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO billing_sessions (id, account_id, payment_method, amount_micro_usd, external_id, status, referral_code)
-		 VALUES ($1, $2, $3, $4, $5, $6, $7)`,
-		session.ID, session.AccountID, session.PaymentMethod,
-		session.AmountMicroUSD, session.ExternalID, session.Status, session.ReferralCode,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create billing session: %w", err)
-	}
-	return nil
-}
-
-// GetBillingSession retrieves a billing session by ID.
-func (s *PostgresStore) GetBillingSession(sessionID string) (*BillingSession, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var bs BillingSession
-	err := s.pool.QueryRow(ctx,
-		`SELECT id, account_id, payment_method, amount_micro_usd, external_id, status, referral_code, created_at, completed_at
-		 FROM billing_sessions WHERE id = $1`, sessionID,
-	).Scan(&bs.ID, &bs.AccountID, &bs.PaymentMethod,
-		&bs.AmountMicroUSD, &bs.ExternalID, &bs.Status, &bs.ReferralCode,
-		&bs.CreatedAt, &bs.CompletedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: billing session not found: %w", err)
-	}
-	return &bs, nil
-}
-
-// CompleteBillingSession marks a session as completed.
-func (s *PostgresStore) CompleteBillingSession(sessionID string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE billing_sessions SET status = 'completed', completed_at = NOW()
-		 WHERE id = $1 AND status = 'pending'`, sessionID,
-	)
-	if err != nil {
-		return fmt.Errorf("store: complete billing session: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("store: billing session %q not found or already completed", sessionID)
-	}
-	return nil
-}
-
-// IsExternalIDProcessed returns true if a completed billing session with this external ID exists.
-func (s *PostgresStore) IsExternalIDProcessed(externalID string) bool {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var count int
-	_ = s.pool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM billing_sessions WHERE external_id = $1 AND status = 'completed'`,
-		externalID,
-	).Scan(&count)
-	return count > 0
-}
-
-// --- Custom Pricing ---
-
-func (s *PostgresStore) SetModelPrice(accountID, model string, inputPrice, outputPrice int64) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO model_prices (account_id, model, input_price, output_price, updated_at)
-		 VALUES ($1, $2, $3, $4, NOW())
-		 ON CONFLICT (account_id, model) DO UPDATE SET
-		   input_price = $3, output_price = $4, updated_at = NOW()`,
-		accountID, model, inputPrice, outputPrice,
-	)
-	if err != nil {
-		return fmt.Errorf("store: set model price: %w", err)
-	}
-
-	// Invalidate cache.
-	key := accountID + ":" + model
-	s.priceCacheMu.Lock()
-	delete(s.priceCache, key)
-	s.priceCacheMu.Unlock()
-
-	return nil
-}
-
-func (s *PostgresStore) GetModelPrice(accountID, model string) (int64, int64, bool) {
-	key := accountID + ":" + model
-
-	// Check in-memory cache (30-second TTL).
-	s.priceCacheMu.RLock()
-	if cached, ok := s.priceCache[key]; ok && time.Since(cached.at) < 30*time.Second {
-		s.priceCacheMu.RUnlock()
-		return cached.input, cached.output, true
-	}
-	s.priceCacheMu.RUnlock()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var input, output int64
-	err := s.pool.QueryRow(ctx,
-		`SELECT input_price, output_price FROM model_prices WHERE account_id = $1 AND model = $2`,
-		accountID, model,
-	).Scan(&input, &output)
-	if err != nil {
-		return 0, 0, false
-	}
-
-	// Populate cache.
-	s.priceCacheMu.Lock()
-	s.priceCache[key] = cachedPrice{input: input, output: output, at: time.Now()}
-	s.priceCacheMu.Unlock()
-
-	return input, output, true
-}
-
-func (s *PostgresStore) ListModelPrices(accountID string) []ModelPrice {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT account_id, model, input_price, output_price FROM model_prices WHERE account_id = $1 ORDER BY model`,
-		accountID,
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var prices []ModelPrice
-	for rows.Next() {
-		var mp ModelPrice
-		if err := rows.Scan(&mp.AccountID, &mp.Model, &mp.InputPrice, &mp.OutputPrice); err != nil {
-			continue
-		}
-		prices = append(prices, mp)
-	}
-	return prices
-}
-
-func (s *PostgresStore) DeleteModelPrice(accountID, model string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`DELETE FROM model_prices WHERE account_id = $1 AND model = $2`,
-		accountID, model,
-	)
-	if err != nil {
-		return fmt.Errorf("store: delete model price: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("no custom price for model %q", model)
-	}
-	return nil
-}
-
-// --- Users (Privy) ---
-
-// CreateUser creates a new user record linked to a Privy identity.
-func (s *PostgresStore) CreateUser(user *User) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO users (account_id, privy_user_id, email, role, platform_fee_percent)
-		 VALUES ($1, $2, $3, $4, $5)`,
-		user.AccountID, user.PrivyUserID, user.Email, user.Role, user.PlatformFeePercent,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create user: %w", err)
-	}
-	return nil
-}
-
-const userSelectColumns = `account_id, privy_user_id, email, role, platform_fee_percent,
-	stripe_account_id, stripe_account_status, stripe_destination_type,
-	stripe_destination_last4, stripe_instant_eligible, created_at`
-
-func scanUser(row interface {
-	Scan(...any) error
-}) (*User, error) {
-	var u User
-	if err := row.Scan(&u.AccountID, &u.PrivyUserID, &u.Email, &u.Role, &u.PlatformFeePercent,
-		&u.StripeAccountID, &u.StripeAccountStatus, &u.StripeDestinationType,
-		&u.StripeDestinationLast4, &u.StripeInstantEligible, &u.CreatedAt); err != nil {
-		return nil, err
-	}
-	return &u, nil
-}
-
-// GetUserByPrivyID returns the user for a Privy DID.
-func (s *PostgresStore) GetUserByPrivyID(privyUserID string) (*User, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+userSelectColumns+` FROM users WHERE privy_user_id = $1`, privyUserID,
-	)
-	u, err := scanUser(row)
-	if err != nil {
-		return nil, fmt.Errorf("store: user not found: %w", err)
-	}
-	return u, nil
-}
-
-// GetUserByAccountID returns the user for an internal account ID.
-func (s *PostgresStore) GetUserByAccountID(accountID string) (*User, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+userSelectColumns+` FROM users WHERE account_id = $1`, accountID,
-	)
-	u, err := scanUser(row)
-	if err != nil {
-		return nil, fmt.Errorf("store: user not found: %w", err)
-	}
-	return u, nil
-}
-
-// SetUserStripeAccount upserts the Stripe Connect fields on a user record.
-func (s *PostgresStore) SetUserStripeAccount(accountID, stripeAccountID, status, destinationType, destinationLast4 string, instantEligible bool) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE users SET
-			stripe_account_id = $2,
-			stripe_account_status = $3,
-			stripe_destination_type = $4,
-			stripe_destination_last4 = $5,
-			stripe_instant_eligible = $6
-		 WHERE account_id = $1`,
-		accountID, stripeAccountID, status, destinationType, destinationLast4, instantEligible,
-	)
-	if err != nil {
-		return fmt.Errorf("store: set stripe account: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("user with account ID %q not found", accountID)
-	}
-	return nil
-}
-
-// GetUserByStripeAccount finds a user by their Stripe connected account ID.
-func (s *PostgresStore) GetUserByStripeAccount(stripeAccountID string) (*User, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+userSelectColumns+` FROM users WHERE stripe_account_id = $1`, stripeAccountID,
-	)
-	u, err := scanUser(row)
-	if err != nil {
-		return nil, fmt.Errorf("store: user with Stripe account %q not found: %w", stripeAccountID, err)
-	}
-	return u, nil
-}
-
-// SetUserRole sets the account role on a user record.
-func (s *PostgresStore) SetUserRole(accountID, role string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE users SET role = $2 WHERE account_id = $1`,
-		accountID, role,
-	)
-	if err != nil {
-		return fmt.Errorf("store: set user role: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("user with account ID %q not found", accountID)
-	}
-	return nil
-}
-
-// SetUserPlatformFeePercent sets (or clears, when nil) the per-account platform
-// fee override.
-func (s *PostgresStore) SetUserPlatformFeePercent(accountID string, feePercent *int64) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE users SET platform_fee_percent = $2 WHERE account_id = $1`,
-		accountID, feePercent,
-	)
-	if err != nil {
-		return fmt.Errorf("store: set user platform fee: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("user with account ID %q not found", accountID)
-	}
-	return nil
-}
-
-// GetUserByEmail returns the user for an email address.
-func (s *PostgresStore) GetUserByEmail(email string) (*User, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+userSelectColumns+` FROM users WHERE LOWER(email) = LOWER($1)`, email,
-	)
-	u, err := scanUser(row)
-	if err != nil {
-		return nil, fmt.Errorf("user with email %q not found", email)
-	}
-	return u, nil
-}
-
-// --- Stripe Withdrawals ---
-
-func (s *PostgresStore) CreateStripeWithdrawal(w *StripeWithdrawal) error {
-	if w == nil || w.ID == "" {
-		return errors.New("stripe withdrawal id is required")
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	now := time.Now()
-	if w.CreatedAt.IsZero() {
-		w.CreatedAt = now
-	}
-	if w.UpdatedAt.IsZero() {
-		w.UpdatedAt = w.CreatedAt
-	}
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO stripe_withdrawals
-		 (id, account_id, stripe_account_id, transfer_id, payout_id,
-		  amount_micro_usd, fee_micro_usd, net_micro_usd, method, status,
-		  failure_reason, refunded, created_at, updated_at)
-		 VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14)`,
-		w.ID, w.AccountID, w.StripeAccountID, w.TransferID, w.PayoutID,
-		w.AmountMicroUSD, w.FeeMicroUSD, w.NetMicroUSD, w.Method, w.Status,
-		w.FailureReason, w.Refunded, w.CreatedAt, w.UpdatedAt,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create stripe withdrawal: %w", err)
-	}
-	return nil
-}
-
-const stripeWithdrawalSelectColumns = `id, account_id, stripe_account_id, transfer_id, payout_id,
-	amount_micro_usd, fee_micro_usd, net_micro_usd, method, status,
-	failure_reason, refunded, created_at, updated_at`
-
-func scanStripeWithdrawal(row interface{ Scan(...any) error }) (*StripeWithdrawal, error) {
-	var w StripeWithdrawal
-	if err := row.Scan(&w.ID, &w.AccountID, &w.StripeAccountID, &w.TransferID, &w.PayoutID,
-		&w.AmountMicroUSD, &w.FeeMicroUSD, &w.NetMicroUSD, &w.Method, &w.Status,
-		&w.FailureReason, &w.Refunded, &w.CreatedAt, &w.UpdatedAt); err != nil {
-		return nil, err
-	}
-	return &w, nil
-}
-
-func (s *PostgresStore) GetStripeWithdrawal(id string) (*StripeWithdrawal, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+stripeWithdrawalSelectColumns+` FROM stripe_withdrawals WHERE id = $1`, id)
-	w, err := scanStripeWithdrawal(row)
-	if err != nil {
-		return nil, fmt.Errorf("store: stripe withdrawal %q not found: %w", id, err)
-	}
-	return w, nil
-}
-
-func (s *PostgresStore) GetStripeWithdrawalByPayoutID(payoutID string) (*StripeWithdrawal, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+stripeWithdrawalSelectColumns+` FROM stripe_withdrawals WHERE payout_id = $1`, payoutID)
-	w, err := scanStripeWithdrawal(row)
-	if err != nil {
-		return nil, fmt.Errorf("store: stripe withdrawal with payout %q not found: %w", payoutID, err)
-	}
-	return w, nil
-}
-
-func (s *PostgresStore) GetStripeWithdrawalByTransferID(transferID string) (*StripeWithdrawal, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	row := s.pool.QueryRow(ctx,
-		`SELECT `+stripeWithdrawalSelectColumns+` FROM stripe_withdrawals WHERE transfer_id = $1`, transferID)
-	w, err := scanStripeWithdrawal(row)
-	if err != nil {
-		return nil, fmt.Errorf("store: stripe withdrawal with transfer %q not found: %w", transferID, err)
-	}
-	return w, nil
-}
-
-func (s *PostgresStore) UpdateStripeWithdrawal(w *StripeWithdrawal) error {
-	if w == nil || w.ID == "" {
-		return errors.New("stripe withdrawal id is required")
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE stripe_withdrawals SET
-			transfer_id = $2, payout_id = $3, status = $4,
-			failure_reason = $5, refunded = $6, updated_at = NOW()
-		 WHERE id = $1`,
-		w.ID, w.TransferID, w.PayoutID, w.Status, w.FailureReason, w.Refunded,
-	)
-	if err != nil {
-		return fmt.Errorf("store: update stripe withdrawal: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("stripe withdrawal %q not found", w.ID)
-	}
-	w.UpdatedAt = time.Now()
-	return nil
-}
-
-func (s *PostgresStore) ListStripeWithdrawals(accountID string, limit int) ([]StripeWithdrawal, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	q := `SELECT ` + stripeWithdrawalSelectColumns + ` FROM stripe_withdrawals WHERE account_id = $1 ORDER BY created_at DESC`
-	args := []any{accountID}
-	if limit > 0 {
-		q += ` LIMIT $2`
-		args = append(args, limit)
-	}
-	rows, err := s.pool.Query(ctx, q, args...)
-	if err != nil {
-		return nil, fmt.Errorf("store: list stripe withdrawals: %w", err)
-	}
-	defer rows.Close()
-
-	var out []StripeWithdrawal
-	for rows.Next() {
-		w, err := scanStripeWithdrawal(rows)
-		if err != nil {
-			return nil, fmt.Errorf("store: scan stripe withdrawal: %w", err)
-		}
-		out = append(out, *w)
-	}
-	if out == nil {
-		return []StripeWithdrawal{}, nil
-	}
-	return out, nil
-}
-
-// --- Releases ---
-
-func (s *PostgresStore) SetRelease(release *Release) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO releases (version, platform, backend, binary_hash, bundle_hash, metallib_hash, python_hash, runtime_hash, template_hashes, url, changelog, active, created_at)
-		 VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, TRUE, NOW())
-		 ON CONFLICT (version, platform) DO UPDATE SET
-		   backend = $3, binary_hash = $4, bundle_hash = $5, metallib_hash = $6, python_hash = $7, runtime_hash = $8, template_hashes = $9, url = $10, changelog = $11, active = TRUE`,
-		release.Version, release.Platform, release.Backend, release.BinaryHash, release.BundleHash,
-		release.MetallibHash, release.PythonHash, release.RuntimeHash, release.TemplateHashes,
-		release.URL, release.Changelog,
-	)
-	if err != nil {
-		return fmt.Errorf("store: set release: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) ListReleases() []Release {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT version, platform, COALESCE(backend, ''), binary_hash, bundle_hash, COALESCE(metallib_hash, ''),
-		        COALESCE(python_hash, ''), COALESCE(runtime_hash, ''), COALESCE(template_hashes, ''),
-		        url, changelog, active, created_at
-		 FROM releases ORDER BY created_at DESC`,
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var releases []Release
-	for rows.Next() {
-		var r Release
-		if err := rows.Scan(&r.Version, &r.Platform, &r.Backend, &r.BinaryHash, &r.BundleHash, &r.MetallibHash,
-			&r.PythonHash, &r.RuntimeHash, &r.TemplateHashes,
-			&r.URL, &r.Changelog, &r.Active, &r.CreatedAt); err != nil {
-			continue
-		}
-		releases = append(releases, r)
-	}
-	return releases
-}
-
-func (s *PostgresStore) GetLatestRelease(platform string) *Release {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT version, platform, COALESCE(backend, ''), binary_hash, bundle_hash, COALESCE(metallib_hash, ''),
-		        COALESCE(python_hash, ''), COALESCE(runtime_hash, ''), COALESCE(template_hashes, ''),
-		        url, changelog, active, created_at
-		 FROM releases WHERE platform = $1 AND active = TRUE`, platform,
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var latest *Release
-	for rows.Next() {
-		var r Release
-		if err := rows.Scan(&r.Version, &r.Platform, &r.Backend, &r.BinaryHash, &r.BundleHash, &r.MetallibHash,
-			&r.PythonHash, &r.RuntimeHash, &r.TemplateHashes,
-			&r.URL, &r.Changelog, &r.Active, &r.CreatedAt); err != nil {
-			return nil
-		}
-		if latest == nil ||
-			releaseVersionGreater(r.Version, latest.Version) ||
-			(r.Version == latest.Version && r.CreatedAt.After(latest.CreatedAt)) {
-			copy := r
-			latest = &copy
-		}
-	}
-	if rows.Err() != nil || latest == nil {
-		return nil
-	}
-	return latest
-}
-
-func (s *PostgresStore) DeleteRelease(version, platform string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE releases SET active = FALSE WHERE version = $1 AND platform = $2`,
-		version, platform,
-	)
-	if err != nil {
-		return fmt.Errorf("store: delete release: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("release %s/%s not found", version, platform)
-	}
-	return nil
-}
-
-// --- Device Authorization ---
-
-func (s *PostgresStore) CreateDeviceCode(dc *DeviceCode) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO device_codes (device_code, user_code, account_id, status, expires_at)
-		 VALUES ($1, $2, $3, $4, $5)`,
-		dc.DeviceCode, dc.UserCode, dc.AccountID, dc.Status, dc.ExpiresAt,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create device code: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) GetDeviceCode(deviceCode string) (*DeviceCode, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var dc DeviceCode
-	err := s.pool.QueryRow(ctx,
-		`SELECT device_code, user_code, account_id, status, expires_at, created_at
-		 FROM device_codes WHERE device_code = $1`, deviceCode,
-	).Scan(&dc.DeviceCode, &dc.UserCode, &dc.AccountID, &dc.Status, &dc.ExpiresAt, &dc.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: device code not found: %w", err)
-	}
-	return &dc, nil
-}
-
-func (s *PostgresStore) GetDeviceCodeByUserCode(userCode string) (*DeviceCode, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var dc DeviceCode
-	err := s.pool.QueryRow(ctx,
-		`SELECT device_code, user_code, account_id, status, expires_at, created_at
-		 FROM device_codes WHERE user_code = $1`, userCode,
-	).Scan(&dc.DeviceCode, &dc.UserCode, &dc.AccountID, &dc.Status, &dc.ExpiresAt, &dc.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: user code not found: %w", err)
-	}
-	return &dc, nil
-}
-
-func (s *PostgresStore) ApproveDeviceCode(deviceCode, accountID string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE device_codes SET status = 'approved', account_id = $2
-		 WHERE device_code = $1 AND status = 'pending' AND expires_at > NOW()`,
-		deviceCode, accountID,
-	)
-	if err != nil {
-		return fmt.Errorf("store: approve device code: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return errors.New("device code not found, not pending, or expired")
-	}
-	return nil
-}
-
-func (s *PostgresStore) DeleteExpiredDeviceCodes() error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx, `DELETE FROM device_codes WHERE expires_at < NOW()`)
-	if err != nil {
-		return fmt.Errorf("store: delete expired device codes: %w", err)
-	}
-	return nil
-}
-
-// --- Provider Tokens ---
-
-func (s *PostgresStore) CreateProviderToken(pt *ProviderToken) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO provider_tokens (token_hash, account_id, label, active)
-		 VALUES ($1, $2, $3, $4)`,
-		pt.TokenHash, pt.AccountID, pt.Label, pt.Active,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create provider token: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) GetProviderToken(token string) (*ProviderToken, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	h := HashKey(token)
-	var pt ProviderToken
-	err := s.pool.QueryRow(ctx,
-		`SELECT token_hash, account_id, label, active, created_at
-		 FROM provider_tokens WHERE token_hash = $1 AND active = TRUE`, h,
-	).Scan(&pt.TokenHash, &pt.AccountID, &pt.Label, &pt.Active, &pt.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: provider token not found: %w", err)
-	}
-	return &pt, nil
-}
-
-func (s *PostgresStore) RevokeProviderToken(token string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	h := HashKey(token)
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE provider_tokens SET active = FALSE WHERE token_hash = $1`, h,
-	)
-	if err != nil {
-		return fmt.Errorf("store: revoke provider token: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return errors.New("provider token not found")
-	}
-	return nil
-}
-
-// --- Invite Codes ---
-
-func (s *PostgresStore) CreateInviteCode(code *InviteCode) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO invite_codes (code, amount_micro_usd, max_uses, used_count, active, expires_at)
-		 VALUES ($1, $2, $3, $4, $5, $6)`,
-		code.Code, code.AmountMicroUSD, code.MaxUses, code.UsedCount, code.Active, code.ExpiresAt,
-	)
-	if err != nil {
-		return fmt.Errorf("store: create invite code: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) GetInviteCode(code string) (*InviteCode, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var ic InviteCode
-	err := s.pool.QueryRow(ctx,
-		`SELECT code, amount_micro_usd, max_uses, used_count, active, expires_at, created_at
-		 FROM invite_codes WHERE code = $1`, code,
-	).Scan(&ic.Code, &ic.AmountMicroUSD, &ic.MaxUses, &ic.UsedCount, &ic.Active, &ic.ExpiresAt, &ic.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: invite code not found: %w", err)
-	}
-	return &ic, nil
-}
-
-func (s *PostgresStore) ListInviteCodes() []InviteCode {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT code, amount_micro_usd, max_uses, used_count, active, expires_at, created_at
-		 FROM invite_codes ORDER BY created_at DESC`,
-	)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var codes []InviteCode
-	for rows.Next() {
-		var ic InviteCode
-		if err := rows.Scan(&ic.Code, &ic.AmountMicroUSD, &ic.MaxUses, &ic.UsedCount, &ic.Active, &ic.ExpiresAt, &ic.CreatedAt); err != nil {
-			continue
-		}
-		codes = append(codes, ic)
-	}
-	return codes
-}
-
-func (s *PostgresStore) DeactivateInviteCode(code string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE invite_codes SET active = FALSE WHERE code = $1`, code,
-	)
-	if err != nil {
-		return fmt.Errorf("store: deactivate invite code: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("invite code %q not found", code)
-	}
-	return nil
-}
-
-func (s *PostgresStore) RedeemInviteCode(code string, accountID string) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("store: begin tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	// Lock the invite code row
-	var ic InviteCode
-	err = tx.QueryRow(ctx,
-		`SELECT code, amount_micro_usd, max_uses, used_count, active, expires_at
-		 FROM invite_codes WHERE code = $1 FOR UPDATE`, code,
-	).Scan(&ic.Code, &ic.AmountMicroUSD, &ic.MaxUses, &ic.UsedCount, &ic.Active, &ic.ExpiresAt)
-	if err != nil {
-		return fmt.Errorf("invite code %q not found", code)
-	}
-	if !ic.Active {
-		return fmt.Errorf("invite code %q is inactive", code)
-	}
-	if ic.ExpiresAt != nil && time.Now().After(*ic.ExpiresAt) {
-		return fmt.Errorf("invite code %q has expired", code)
-	}
-	if ic.MaxUses > 0 && ic.UsedCount >= ic.MaxUses {
-		return fmt.Errorf("invite code %q has reached max uses", code)
-	}
-
-	// Insert redemption (PK constraint prevents double-redemption)
-	_, err = tx.Exec(ctx,
-		`INSERT INTO invite_redemptions (code, account_id) VALUES ($1, $2)`,
-		code, accountID,
-	)
-	if err != nil {
-		return fmt.Errorf("account has already redeemed code %q", code)
-	}
-
-	// Increment used_count
-	_, err = tx.Exec(ctx,
-		`UPDATE invite_codes SET used_count = used_count + 1 WHERE code = $1`, code,
-	)
-	if err != nil {
-		return fmt.Errorf("store: update invite code: %w", err)
-	}
-
-	return tx.Commit(ctx)
-}
-
-func (s *PostgresStore) HasRedeemedInviteCode(code, accountID string) bool {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var count int
-	_ = s.pool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM invite_redemptions WHERE code = $1 AND account_id = $2`,
-		code, accountID,
-	).Scan(&count)
-	return count > 0
-}
-
-// --- Provider Earnings ---
-
-// RecordProviderEarning stores an earning record for a specific provider node.
-func (s *PostgresStore) RecordProviderEarning(earning *ProviderEarning) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO provider_earnings (account_id, provider_id, provider_key, job_id, model, amount_micro_usd, prompt_tokens, completion_tokens)
-		 VALUES ($1, $2, $3, $4, $5, $6, $7, $8)`,
-		earning.AccountID, earning.ProviderID, earning.ProviderKey, earning.JobID,
-		earning.Model, earning.AmountMicroUSD, earning.PromptTokens, earning.CompletionTokens,
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert provider earning: %w", err)
-	}
-	return nil
-}
-
-// GetProviderEarnings returns earnings for a specific provider node (by public key), newest first.
-func (s *PostgresStore) GetProviderEarnings(providerKey string, limit int) ([]ProviderEarning, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT id, account_id, provider_id, provider_key, job_id, model, amount_micro_usd, prompt_tokens, completion_tokens, created_at
-		 FROM provider_earnings
-		 WHERE provider_key = $1
-		 ORDER BY created_at DESC
-		 LIMIT $2`,
-		providerKey, limit,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: query provider earnings: %w", err)
-	}
-	defer rows.Close()
-
-	var results []ProviderEarning
-	for rows.Next() {
-		var e ProviderEarning
-		if err := rows.Scan(&e.ID, &e.AccountID, &e.ProviderID, &e.ProviderKey, &e.JobID,
-			&e.Model, &e.AmountMicroUSD, &e.PromptTokens, &e.CompletionTokens, &e.CreatedAt); err != nil {
-			continue
-		}
-		results = append(results, e)
-	}
-	if results == nil {
-		return []ProviderEarning{}, nil
-	}
-	return results, nil
-}
-
-// GetAccountEarnings returns all earnings across all nodes for an account, newest first.
-func (s *PostgresStore) GetAccountEarnings(accountID string, limit int) ([]ProviderEarning, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT id, account_id, provider_id, provider_key, job_id, model, amount_micro_usd, prompt_tokens, completion_tokens, created_at
-		 FROM provider_earnings
-		 WHERE account_id = $1
-		 ORDER BY created_at DESC
-		 LIMIT $2`,
-		accountID, limit,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: query account earnings: %w", err)
-	}
-	defer rows.Close()
-
-	var results []ProviderEarning
-	for rows.Next() {
-		var e ProviderEarning
-		if err := rows.Scan(&e.ID, &e.AccountID, &e.ProviderID, &e.ProviderKey, &e.JobID,
-			&e.Model, &e.AmountMicroUSD, &e.PromptTokens, &e.CompletionTokens, &e.CreatedAt); err != nil {
-			continue
-		}
-		results = append(results, e)
-	}
-	if results == nil {
-		return []ProviderEarning{}, nil
-	}
-	return results, nil
-}
-
-// GetProviderEarningsSummary returns lifetime aggregates for a provider node.
-// Reads from the materialized earnings_summary table (PK lookup) instead of
-// scanning all provider_earnings rows.
-func (s *PostgresStore) GetProviderEarningsSummary(providerKey string) (ProviderEarningsSummary, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var summary ProviderEarningsSummary
-	err := s.pool.QueryRow(ctx,
-		`SELECT total_count, total_micro_usd, total_prompt_tokens, total_completion_tokens
-		 FROM earnings_summary
-		 WHERE key = $1 AND key_type = 'provider'`,
-		providerKey,
-	).Scan(&summary.Count, &summary.TotalMicroUSD, &summary.PromptTokens, &summary.CompletionTokens)
-	if err != nil {
-		// No rows = no earnings yet, return zeros (not an error).
-		return ProviderEarningsSummary{}, nil
-	}
-
-	return summary, nil
-}
-
-// GetAccountEarningsSummary returns lifetime aggregates for an account.
-// Reads from the materialized earnings_summary table (PK lookup) instead of
-// scanning all provider_earnings rows.
-func (s *PostgresStore) GetAccountEarningsSummary(accountID string) (ProviderEarningsSummary, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var summary ProviderEarningsSummary
-	err := s.pool.QueryRow(ctx,
-		`SELECT total_count, total_micro_usd, total_prompt_tokens, total_completion_tokens
-		 FROM earnings_summary
-		 WHERE key = $1 AND key_type = 'account'`,
-		accountID,
-	).Scan(&summary.Count, &summary.TotalMicroUSD, &summary.PromptTokens, &summary.CompletionTokens)
-	if err != nil {
-		// No rows = no earnings yet, return zeros (not an error).
-		return ProviderEarningsSummary{}, nil
-	}
-
-	return summary, nil
-}
-
-// RecordProviderPayout stores a payout record for a provider wallet.
-func (s *PostgresStore) RecordProviderPayout(payout *ProviderPayout) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO provider_payouts (provider_address, amount_micro_usd, model, job_id, settled, created_at)
-		 VALUES ($1, $2, $3, $4, $5, COALESCE($6, NOW()))`,
-		payout.ProviderAddress, payout.AmountMicroUSD, payout.Model, payout.JobID, payout.Settled, nullableCreatedAt(payout.Timestamp),
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert provider payout: %w", err)
-	}
-
-	return nil
-}
-
-// ListProviderPayouts returns all provider payout records in creation order.
-func (s *PostgresStore) ListProviderPayouts() ([]ProviderPayout, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT id, provider_address, amount_micro_usd, model, job_id, settled, created_at
-		 FROM provider_payouts
-		 ORDER BY id ASC`,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: query provider payouts: %w", err)
-	}
-	defer rows.Close()
-
-	var results []ProviderPayout
-	for rows.Next() {
-		var payout ProviderPayout
-		if err := rows.Scan(&payout.ID, &payout.ProviderAddress, &payout.AmountMicroUSD, &payout.Model, &payout.JobID, &payout.Settled, &payout.Timestamp); err != nil {
-			continue
-		}
-		results = append(results, payout)
-	}
-	if results == nil {
-		return []ProviderPayout{}, nil
-	}
-
-	return results, nil
-}
-
-// SettleProviderPayout marks a provider payout as settled.
-func (s *PostgresStore) SettleProviderPayout(id int64) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tag, err := s.pool.Exec(ctx,
-		`UPDATE provider_payouts
-		 SET settled = TRUE
-		 WHERE id = $1 AND settled = FALSE`,
-		id,
-	)
-	if err != nil {
-		return fmt.Errorf("store: settle provider payout: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return fmt.Errorf("provider payout %d not found or already settled", id)
-	}
-
-	return nil
-}
-
-// CreditProviderAccount atomically credits a linked provider account and records
-// the corresponding per-node earning.
-//
-// Single-statement CTE: upsert balance, insert ledger entry, insert earning --
-// all in one round trip. The old implementation used 6 sequential round trips
-// (BEGIN + upsert + SELECT balance + INSERT ledger + INSERT earning + COMMIT).
-func (s *PostgresStore) CreditProviderAccount(earning *ProviderEarning) error {
-	if earning == nil {
-		return errors.New("provider earning is required")
-	}
-	if earning.AccountID == "" {
-		return errors.New("provider earning account_id is required")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var balanceAfter int64
-	err := s.pool.QueryRow(ctx, `
-		WITH credit AS (
-			INSERT INTO balances (account_id, balance_micro_usd, withdrawable_micro_usd, updated_at)
-			VALUES ($1, $2, $2, NOW())
-			ON CONFLICT (account_id) DO UPDATE SET
-			  balance_micro_usd = balances.balance_micro_usd + $2,
-			  withdrawable_micro_usd = balances.withdrawable_micro_usd + $2,
-			  updated_at = NOW()
-			RETURNING balance_micro_usd
-		), ledger AS (
-			INSERT INTO ledger_entries (account_id, entry_type, amount_micro_usd, balance_after, reference, created_at)
-			SELECT $1, $3, $2, balance_micro_usd, $4, COALESCE($5::timestamptz, NOW())
-			FROM credit
-		), earning AS (
-			INSERT INTO provider_earnings (
-				account_id, provider_id, provider_key, job_id, model, amount_micro_usd, prompt_tokens, completion_tokens, created_at
-			) VALUES ($1, $6, $7, $4, $8, $2, $9, $10, COALESCE($5::timestamptz, NOW()))
-		), summary_account AS (
-			INSERT INTO earnings_summary (key, key_type, total_count, total_micro_usd, total_prompt_tokens, total_completion_tokens, updated_at)
-			VALUES ($1, 'account', 1, $2, $9, $10, NOW())
-			ON CONFLICT (key, key_type) DO UPDATE SET
-			  total_count = earnings_summary.total_count + 1,
-			  total_micro_usd = earnings_summary.total_micro_usd + $2,
-			  total_prompt_tokens = earnings_summary.total_prompt_tokens + $9,
-			  total_completion_tokens = earnings_summary.total_completion_tokens + $10,
-			  updated_at = NOW()
-		), summary_provider AS (
-			INSERT INTO earnings_summary (key, key_type, total_count, total_micro_usd, total_prompt_tokens, total_completion_tokens, updated_at)
-			VALUES ($7, 'provider', 1, $2, $9, $10, NOW())
-			ON CONFLICT (key, key_type) DO UPDATE SET
-			  total_count = earnings_summary.total_count + 1,
-			  total_micro_usd = earnings_summary.total_micro_usd + $2,
-			  total_prompt_tokens = earnings_summary.total_prompt_tokens + $9,
-			  total_completion_tokens = earnings_summary.total_completion_tokens + $10,
-			  updated_at = NOW()
-		)
-		SELECT balance_micro_usd FROM credit`,
-		earning.AccountID,                    // $1
-		earning.AmountMicroUSD,               // $2
-		string(LedgerPayout),                 // $3
-		earning.JobID,                        // $4
-		nullableCreatedAt(earning.CreatedAt), // $5
-		earning.ProviderID,                   // $6
-		earning.ProviderKey,                  // $7
-		earning.Model,                        // $8
-		earning.PromptTokens,                 // $9
-		earning.CompletionTokens,             // $10
-	).Scan(&balanceAfter)
-	if err != nil {
-		return fmt.Errorf("store: credit provider account: %w", err)
-	}
-	return nil
-}
-
-// CreditProviderWallet atomically credits an unlinked provider wallet and
-// records the corresponding payout history row.
-func (s *PostgresStore) CreditProviderWallet(payout *ProviderPayout) error {
-	if payout == nil {
-		return errors.New("provider payout is required")
-	}
-	if payout.ProviderAddress == "" {
-		return errors.New("provider payout address is required")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("store: begin tx: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	if err := creditWithdrawableTx(ctx, tx, payout.ProviderAddress, payout.AmountMicroUSD, LedgerPayout, payout.JobID, payout.Timestamp); err != nil {
-		return err
-	}
-
-	_, err = tx.Exec(ctx,
-		`INSERT INTO provider_payouts (provider_address, amount_micro_usd, model, job_id, settled, created_at)
-		 VALUES ($1, $2, $3, $4, $5, COALESCE($6, NOW()))`,
-		payout.ProviderAddress,
-		payout.AmountMicroUSD,
-		payout.Model,
-		payout.JobID,
-		payout.Settled,
-		nullableCreatedAt(payout.Timestamp),
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert provider payout: %w", err)
-	}
-
-	return tx.Commit(ctx)
-}
-
-// --- Provider Fleet Persistence ---
-
-func marshalProviderLocation(loc *ProviderLocation) json.RawMessage {
-	if loc == nil {
-		return nil
-	}
-	b, err := json.Marshal(loc)
-	if err != nil {
-		return nil
-	}
-	return b
-}
-
-func unmarshalProviderLocation(raw []byte) *ProviderLocation {
-	if len(raw) == 0 {
-		return nil
-	}
-	var loc ProviderLocation
-	if err := json.Unmarshal(raw, &loc); err != nil {
-		return nil
-	}
-	return &loc
-}
-
-func (s *PostgresStore) UpsertProvider(ctx context.Context, p ProviderRecord) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO providers (
-			id, hardware, models, backend, location, trust_level, attested,
-			attestation_result, se_public_key, serial_number,
-			mda_verified, mda_cert_chain, acme_verified,
-			version, runtime_verified, python_hash, runtime_hash,
-			last_challenge_verified, failed_challenges, account_id,
-			lifetime_requests_served, lifetime_tokens_generated,
-			last_session_requests_served, last_session_tokens_generated,
-			registered_at, last_seen
-		) VALUES (
-			$1, $2, $3, $4, $5, $6, $7,
-			$8, $9, $10,
-			$11, $12, $13,
-			$14, $15, $16, $17,
-			$18, $19, $20,
-			$21, $22, $23, $24,
-			$25, $26
-		)
-		ON CONFLICT (id) DO UPDATE SET
-			hardware = $2, models = $3, backend = $4, location = $5,
-			trust_level = $6, attested = $7,
-			attestation_result = $8, se_public_key = $9, serial_number = $10,
-			mda_verified = $11, mda_cert_chain = $12, acme_verified = $13,
-			version = $14, runtime_verified = $15, python_hash = $16, runtime_hash = $17,
-			last_challenge_verified = $18, failed_challenges = $19, account_id = $20,
-			lifetime_requests_served = $21, lifetime_tokens_generated = $22,
-			last_session_requests_served = $23, last_session_tokens_generated = $24,
-			last_seen = $26`,
-		p.ID, p.Hardware, p.Models, p.Backend,
-		marshalProviderLocation(p.Location),
-		p.TrustLevel, p.Attested,
-		p.AttestationResult, p.SEPublicKey, p.SerialNumber,
-		p.MDAVerified, p.MDACertChain, p.ACMEVerified,
-		p.Version, p.RuntimeVerified, p.PythonHash, p.RuntimeHash,
-		p.LastChallengeVerified, p.FailedChallenges, p.AccountID,
-		p.LifetimeRequestsServed, p.LifetimeTokensGenerated,
-		p.LastSessionRequestsServed, p.LastSessionTokensGenerated,
-		p.RegisteredAt, p.LastSeen,
-	)
-	if err != nil {
-		return fmt.Errorf("store: upsert provider: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) GetProviderRecord(ctx context.Context, id string) (*ProviderRecord, error) {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	var p ProviderRecord
-	var locationRaw []byte
-	err := s.pool.QueryRow(ctx,
-		`SELECT id, hardware, models, backend, location, trust_level, attested,
-			attestation_result, se_public_key, serial_number,
-			mda_verified, mda_cert_chain, acme_verified,
-			version, runtime_verified, python_hash, runtime_hash,
-			last_challenge_verified, failed_challenges, account_id,
-			lifetime_requests_served, lifetime_tokens_generated,
-			last_session_requests_served, last_session_tokens_generated,
-			registered_at, last_seen
-		 FROM providers WHERE id = $1`, id,
-	).Scan(
-		&p.ID, &p.Hardware, &p.Models, &p.Backend,
-		&locationRaw,
-		&p.TrustLevel, &p.Attested,
-		&p.AttestationResult, &p.SEPublicKey, &p.SerialNumber,
-		&p.MDAVerified, &p.MDACertChain, &p.ACMEVerified,
-		&p.Version, &p.RuntimeVerified, &p.PythonHash, &p.RuntimeHash,
-		&p.LastChallengeVerified, &p.FailedChallenges, &p.AccountID,
-		&p.LifetimeRequestsServed, &p.LifetimeTokensGenerated,
-		&p.LastSessionRequestsServed, &p.LastSessionTokensGenerated,
-		&p.RegisteredAt, &p.LastSeen,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: provider not found: %w", err)
-	}
-	p.Location = unmarshalProviderLocation(locationRaw)
-	return &p, nil
-}
-
-func (s *PostgresStore) GetProviderBySerial(ctx context.Context, serial string) (*ProviderRecord, error) {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	var p ProviderRecord
-	var locationRaw []byte
-	err := s.pool.QueryRow(ctx,
-		`SELECT id, hardware, models, backend, location, trust_level, attested,
-			attestation_result, se_public_key, serial_number,
-			mda_verified, mda_cert_chain, acme_verified,
-			version, runtime_verified, python_hash, runtime_hash,
-			last_challenge_verified, failed_challenges, account_id,
-			lifetime_requests_served, lifetime_tokens_generated,
-			last_session_requests_served, last_session_tokens_generated,
-			registered_at, last_seen
-		 FROM providers WHERE serial_number = $1 AND serial_number != ''
-		 ORDER BY last_seen DESC LIMIT 1`, serial,
-	).Scan(
-		&p.ID, &p.Hardware, &p.Models, &p.Backend,
-		&locationRaw,
-		&p.TrustLevel, &p.Attested,
-		&p.AttestationResult, &p.SEPublicKey, &p.SerialNumber,
-		&p.MDAVerified, &p.MDACertChain, &p.ACMEVerified,
-		&p.Version, &p.RuntimeVerified, &p.PythonHash, &p.RuntimeHash,
-		&p.LastChallengeVerified, &p.FailedChallenges, &p.AccountID,
-		&p.LifetimeRequestsServed, &p.LifetimeTokensGenerated,
-		&p.LastSessionRequestsServed, &p.LastSessionTokensGenerated,
-		&p.RegisteredAt, &p.LastSeen,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: provider with serial not found: %w", err)
-	}
-	p.Location = unmarshalProviderLocation(locationRaw)
-	return &p, nil
-}
-
-func (s *PostgresStore) ListProviderRecords(ctx context.Context) ([]ProviderRecord, error) {
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT id, hardware, models, backend, location, trust_level, attested,
-			attestation_result, se_public_key, serial_number,
-			mda_verified, mda_cert_chain, acme_verified,
-			version, runtime_verified, python_hash, runtime_hash,
-			last_challenge_verified, failed_challenges, account_id,
-			lifetime_requests_served, lifetime_tokens_generated,
-			last_session_requests_served, last_session_tokens_generated,
-			registered_at, last_seen
-		 FROM providers ORDER BY last_seen DESC`,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: list providers: %w", err)
-	}
-	defer rows.Close()
-
-	var records []ProviderRecord
-	for rows.Next() {
-		var p ProviderRecord
-		var locationRaw []byte
-		if err := rows.Scan(
-			&p.ID, &p.Hardware, &p.Models, &p.Backend,
-			&locationRaw,
-			&p.TrustLevel, &p.Attested,
-			&p.AttestationResult, &p.SEPublicKey, &p.SerialNumber,
-			&p.MDAVerified, &p.MDACertChain, &p.ACMEVerified,
-			&p.Version, &p.RuntimeVerified, &p.PythonHash, &p.RuntimeHash,
-			&p.LastChallengeVerified, &p.FailedChallenges, &p.AccountID,
-			&p.LifetimeRequestsServed, &p.LifetimeTokensGenerated,
-			&p.LastSessionRequestsServed, &p.LastSessionTokensGenerated,
-			&p.RegisteredAt, &p.LastSeen,
-		); err != nil {
-			continue
-		}
-		p.Location = unmarshalProviderLocation(locationRaw)
-		records = append(records, p)
-	}
-	if records == nil {
-		return []ProviderRecord{}, nil
-	}
-	return records, nil
-}
-
-func (s *PostgresStore) ListProvidersByAccount(ctx context.Context, accountID string) ([]ProviderRecord, error) {
-	if accountID == "" {
-		return []ProviderRecord{}, nil
-	}
-
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	// Dedupe in SQL: many session UUIDs can map to the same physical
-	// machine (one row per reconnect). Pick the most-recent row per
-	// stable identity (serial → SE key → id) so we don't return tens
-	// of thousands of historical rows for accounts with churny providers.
-	rows, err := s.pool.Query(ctx,
-		`SELECT DISTINCT ON (
-			COALESCE(NULLIF(serial_number, ''),
-			         NULLIF(se_public_key, ''),
-			         id)
-		 )
-		 id, hardware, models, backend, location, trust_level, attested,
-			attestation_result, se_public_key, serial_number,
-			mda_verified, mda_cert_chain, acme_verified,
-			version, runtime_verified, python_hash, runtime_hash,
-			last_challenge_verified, failed_challenges, account_id,
-			lifetime_requests_served, lifetime_tokens_generated,
-			last_session_requests_served, last_session_tokens_generated,
-			registered_at, last_seen
-		 FROM providers
-		 WHERE account_id = $1
-		 ORDER BY COALESCE(NULLIF(serial_number, ''),
-		                   NULLIF(se_public_key, ''),
-		                   id),
-		          last_seen DESC`,
-		accountID,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: list providers by account: %w", err)
-	}
-	defer rows.Close()
-
-	records := make([]ProviderRecord, 0)
-	for rows.Next() {
-		var p ProviderRecord
-		var locationRaw []byte
-		if err := rows.Scan(
-			&p.ID, &p.Hardware, &p.Models, &p.Backend,
-			&locationRaw,
-			&p.TrustLevel, &p.Attested,
-			&p.AttestationResult, &p.SEPublicKey, &p.SerialNumber,
-			&p.MDAVerified, &p.MDACertChain, &p.ACMEVerified,
-			&p.Version, &p.RuntimeVerified, &p.PythonHash, &p.RuntimeHash,
-			&p.LastChallengeVerified, &p.FailedChallenges, &p.AccountID,
-			&p.LifetimeRequestsServed, &p.LifetimeTokensGenerated,
-			&p.LastSessionRequestsServed, &p.LastSessionTokensGenerated,
-			&p.RegisteredAt, &p.LastSeen,
-		); err != nil {
-			continue
-		}
-		p.Location = unmarshalProviderLocation(locationRaw)
-		records = append(records, p)
-	}
-	return records, nil
-}
-
-func (s *PostgresStore) UpdateProviderLastSeen(ctx context.Context, id string) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`UPDATE providers SET last_seen = NOW() WHERE id = $1`, id,
-	)
-	if err != nil {
-		return fmt.Errorf("store: update provider last_seen: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) UpdateProviderTrust(ctx context.Context, id string, trustLevel string, attested bool, attestationResult json.RawMessage) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`UPDATE providers SET trust_level = $2, attested = $3, attestation_result = $4
-		 WHERE id = $1`,
-		id, trustLevel, attested, attestationResult,
-	)
-	if err != nil {
-		return fmt.Errorf("store: update provider trust: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) UpdateProviderChallenge(ctx context.Context, id string, lastVerified time.Time, failedCount int) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`UPDATE providers SET last_challenge_verified = $2, failed_challenges = $3
-		 WHERE id = $1`,
-		id, lastVerified, failedCount,
-	)
-	if err != nil {
-		return fmt.Errorf("store: update provider challenge: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) UpdateProviderRuntime(ctx context.Context, id string, verified bool, pythonHash, runtimeHash string) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`UPDATE providers SET runtime_verified = $2, python_hash = $3, runtime_hash = $4
-		 WHERE id = $1`,
-		id, verified, pythonHash, runtimeHash,
-	)
-	if err != nil {
-		return fmt.Errorf("store: update provider runtime: %w", err)
-	}
-	return nil
-}
-
-// --- Provider Reputation Persistence ---
-
-func (s *PostgresStore) UpsertReputation(ctx context.Context, providerID string, rep ReputationRecord) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO provider_reputation (
-			provider_id, total_jobs, successful_jobs, failed_jobs,
-			total_uptime_seconds, avg_response_time_ms,
-			challenges_passed, challenges_failed, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, NOW())
-		ON CONFLICT (provider_id) DO UPDATE SET
-			total_jobs = $2, successful_jobs = $3, failed_jobs = $4,
-			total_uptime_seconds = $5, avg_response_time_ms = $6,
-			challenges_passed = $7, challenges_failed = $8,
-			updated_at = NOW()`,
-		providerID, rep.TotalJobs, rep.SuccessfulJobs, rep.FailedJobs,
-		rep.TotalUptimeSeconds, rep.AvgResponseTimeMs,
-		rep.ChallengesPassed, rep.ChallengesFailed,
-	)
-	if err != nil {
-		return fmt.Errorf("store: upsert reputation: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) GetReputation(ctx context.Context, providerID string) (*ReputationRecord, error) {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	var rep ReputationRecord
-	err := s.pool.QueryRow(ctx,
-		`SELECT total_jobs, successful_jobs, failed_jobs,
-			total_uptime_seconds, avg_response_time_ms,
-			challenges_passed, challenges_failed
-		 FROM provider_reputation WHERE provider_id = $1`, providerID,
-	).Scan(
-		&rep.TotalJobs, &rep.SuccessfulJobs, &rep.FailedJobs,
-		&rep.TotalUptimeSeconds, &rep.AvgResponseTimeMs,
-		&rep.ChallengesPassed, &rep.ChallengesFailed,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: reputation not found: %w", err)
-	}
-	return &rep, nil
-}
-
-// --- Provider Log Reports ---
-
-const maxLogReportSize = 10 << 20 // 10 MB
-
-func (s *PostgresStore) StoreLogReport(serialNumber, providerID, accountID string, logData []byte) error {
-	if len(logData) > maxLogReportSize {
-		logData = logData[:maxLogReportSize]
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	_, err := s.pool.Exec(ctx,
-		`INSERT INTO provider_log_reports (serial_number, provider_id, account_id, log_data, log_size_bytes)
-		 VALUES ($1, $2, $3, $4, $5)`,
-		serialNumber, providerID, accountID, logData, int64(len(logData)),
-	)
-	if err != nil {
-		return fmt.Errorf("store: insert log report: %w", err)
-	}
-	return nil
-}
-
-func (s *PostgresStore) GetLogReports(serialNumber string, limit int) ([]LogReport, error) {
-	if limit <= 0 || limit > 100 {
-		limit = 10
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.pool.Query(ctx,
-		`SELECT id, serial_number, provider_id, account_id, log_size_bytes, created_at
-		 FROM provider_log_reports
-		 WHERE serial_number = $1
-		 ORDER BY created_at DESC
-		 LIMIT $2`,
-		serialNumber, limit,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("store: list log reports: %w", err)
-	}
-	defer rows.Close()
-
-	var reports []LogReport
-	for rows.Next() {
-		var r LogReport
-		if err := rows.Scan(&r.ID, &r.SerialNumber, &r.ProviderID, &r.AccountID, &r.LogSizeBytes, &r.CreatedAt); err != nil {
-			continue
-		}
-		reports = append(reports, r)
-	}
-	if reports == nil {
-		return []LogReport{}, nil
-	}
-	return reports, nil
-}
-
-func (s *PostgresStore) GetLogReport(id int64) (*LogReport, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	var r LogReport
-	err := s.pool.QueryRow(ctx,
-		`SELECT id, serial_number, provider_id, account_id, log_data, log_size_bytes, created_at
-		 FROM provider_log_reports WHERE id = $1`, id,
-	).Scan(&r.ID, &r.SerialNumber, &r.ProviderID, &r.AccountID, &r.LogData, &r.LogSizeBytes, &r.CreatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("store: log report %d not found: %w", id, err)
-	}
-	return &r, nil
 }
