+++ coordinator/store/postgres_billing_sessions.go
@@ -1,0 +1,75 @@
+package store
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// --- Billing Sessions ---
+
+// CreateBillingSession stores a new billing session.
+func (s *PostgresStore) CreateBillingSession(session *BillingSession) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := s.pool.Exec(ctx,
+		`INSERT INTO billing_sessions (id, account_id, payment_method, amount_micro_usd, external_id, status, referral_code)
+		 VALUES ($1, $2, $3, $4, $5, $6, $7)`,
+		session.ID, session.AccountID, session.PaymentMethod,
+		session.AmountMicroUSD, session.ExternalID, session.Status, session.ReferralCode,
+	)
+	if err != nil {
+		return fmt.Errorf("store: create billing session: %w", err)
+	}
+	return nil
+}
+
+// GetBillingSession retrieves a billing session by ID.
+func (s *PostgresStore) GetBillingSession(sessionID string) (*BillingSession, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var bs BillingSession
+	err := s.pool.QueryRow(ctx,
+		`SELECT id, account_id, payment_method, amount_micro_usd, external_id, status, referral_code, created_at, completed_at
+		 FROM billing_sessions WHERE id = $1`, sessionID,
+	).Scan(&bs.ID, &bs.AccountID, &bs.PaymentMethod,
+		&bs.AmountMicroUSD, &bs.ExternalID, &bs.Status, &bs.ReferralCode,
+		&bs.CreatedAt, &bs.CompletedAt)
+	if err != nil {
+		return nil, fmt.Errorf("store: billing session not found: %w", err)
+	}
+	return &bs, nil
+}
+
+// CompleteBillingSession marks a session as completed.
+func (s *PostgresStore) CompleteBillingSession(sessionID string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tag, err := s.pool.Exec(ctx,
+		`UPDATE billing_sessions SET status = 'completed', completed_at = NOW()
+		 WHERE id = $1 AND status = 'pending'`, sessionID,
+	)
+	if err != nil {
+		return fmt.Errorf("store: complete billing session: %w", err)
+	}
+	if tag.RowsAffected() == 0 {
+		return fmt.Errorf("store: billing session %q not found or already completed", sessionID)
+	}
+	return nil
+}
+
+// IsExternalIDProcessed returns true if a completed billing session with this external ID exists.
+func (s *PostgresStore) IsExternalIDProcessed(externalID string) bool {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var count int
+	_ = s.pool.QueryRow(ctx,
+		`SELECT COUNT(*) FROM billing_sessions WHERE external_id = $1 AND status = 'completed'`,
+		externalID,
+	).Scan(&count)
+	return count > 0
+}
