+++ coordinator/store/postgres_invites.go
@@ -1,0 +1,140 @@
+package store
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// --- Invite Codes ---
+
+func (s *PostgresStore) CreateInviteCode(code *InviteCode) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := s.pool.Exec(ctx,
+		`INSERT INTO invite_codes (code, amount_micro_usd, max_uses, used_count, active, expires_at)
+		 VALUES ($1, $2, $3, $4, $5, $6)`,
+		code.Code, code.AmountMicroUSD, code.MaxUses, code.UsedCount, code.Active, code.ExpiresAt,
+	)
+	if err != nil {
+		return fmt.Errorf("store: create invite code: %w", err)
+	}
+	return nil
+}
+
+func (s *PostgresStore) GetInviteCode(code string) (*InviteCode, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var ic InviteCode
+	err := s.pool.QueryRow(ctx,
+		`SELECT code, amount_micro_usd, max_uses, used_count, active, expires_at, created_at
+		 FROM invite_codes WHERE code = $1`, code,
+	).Scan(&ic.Code, &ic.AmountMicroUSD, &ic.MaxUses, &ic.UsedCount, &ic.Active, &ic.ExpiresAt, &ic.CreatedAt)
+	if err != nil {
+		return nil, fmt.Errorf("store: invite code not found: %w", err)
+	}
+	return &ic, nil
+}
+
+func (s *PostgresStore) ListInviteCodes() []InviteCode {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	rows, err := s.pool.Query(ctx,
+		`SELECT code, amount_micro_usd, max_uses, used_count, active, expires_at, created_at
+		 FROM invite_codes ORDER BY created_at DESC`,
+	)
+	if err != nil {
+		return nil
+	}
+	defer rows.Close()
+
+	var codes []InviteCode
+	for rows.Next() {
+		var ic InviteCode
+		if err := rows.Scan(&ic.Code, &ic.AmountMicroUSD, &ic.MaxUses, &ic.UsedCount, &ic.Active, &ic.ExpiresAt, &ic.CreatedAt); err != nil {
+			continue
+		}
+		codes = append(codes, ic)
+	}
+	return codes
+}
+
+func (s *PostgresStore) DeactivateInviteCode(code string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tag, err := s.pool.Exec(ctx,
+		`UPDATE invite_codes SET active = FALSE WHERE code = $1`, code,
+	)
+	if err != nil {
+		return fmt.Errorf("store: deactivate invite code: %w", err)
+	}
+	if tag.RowsAffected() == 0 {
+		return fmt.Errorf("invite code %q not found", code)
+	}
+	return nil
+}
+
+func (s *PostgresStore) RedeemInviteCode(code string, accountID string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tx, err := s.pool.Begin(ctx)
+	if err != nil {
+		return fmt.Errorf("store: begin tx: %w", err)
+	}
+	defer tx.Rollback(ctx)
+
+	// Lock the invite code row
+	var ic InviteCode
+	err = tx.QueryRow(ctx,
+		`SELECT code, amount_micro_usd, max_uses, used_count, active, expires_at
+		 FROM invite_codes WHERE code = $1 FOR UPDATE`, code,
+	).Scan(&ic.Code, &ic.AmountMicroUSD, &ic.MaxUses, &ic.UsedCount, &ic.Active, &ic.ExpiresAt)
+	if err != nil {
+		return fmt.Errorf("invite code %q not found", code)
+	}
+	if !ic.Active {
+		return fmt.Errorf("invite code %q is inactive", code)
+	}
+	if ic.ExpiresAt != nil && time.Now().After(*ic.ExpiresAt) {
+		return fmt.Errorf("invite code %q has expired", code)
+	}
+	if ic.MaxUses > 0 && ic.UsedCount >= ic.MaxUses {
+		return fmt.Errorf("invite code %q has reached max uses", code)
+	}
+
+	// Insert redemption (PK constraint prevents double-redemption)
+	_, err = tx.Exec(ctx,
+		`INSERT INTO invite_redemptions (code, account_id) VALUES ($1, $2)`,
+		code, accountID,
+	)
+	if err != nil {
+		return fmt.Errorf("account has already redeemed code %q", code)
+	}
+
+	// Increment used_count
+	_, err = tx.Exec(ctx,
+		`UPDATE invite_codes SET used_count = used_count + 1 WHERE code = $1`, code,
+	)
+	if err != nil {
+		return fmt.Errorf("store: update invite code: %w", err)
+	}
+
+	return tx.Commit(ctx)
+}
+
+func (s *PostgresStore) HasRedeemedInviteCode(code, accountID string) bool {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var count int
+	_ = s.pool.QueryRow(ctx,
+		`SELECT COUNT(*) FROM invite_redemptions WHERE code = $1 AND account_id = $2`,
+		code, accountID,
+	).Scan(&count)
+	return count > 0
+}
