+++ coordinator/store/postgres_referrals.go
@@ -1,0 +1,119 @@
+package store
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// --- Referral System ---
+
+// CreateReferrer registers an account as a referrer with the given code.
+func (s *PostgresStore) CreateReferrer(accountID, code string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := s.pool.Exec(ctx,
+		`INSERT INTO referrers (account_id, code) VALUES ($1, $2)`,
+		accountID, code,
+	)
+	if err != nil {
+		return fmt.Errorf("store: create referrer: %w", err)
+	}
+	return nil
+}
+
+// GetReferrerByCode returns the referrer for a given referral code.
+func (s *PostgresStore) GetReferrerByCode(code string) (*Referrer, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var ref Referrer
+	err := s.pool.QueryRow(ctx,
+		`SELECT account_id, code, created_at FROM referrers WHERE code = $1`, code,
+	).Scan(&ref.AccountID, &ref.Code, &ref.CreatedAt)
+	if err != nil {
+		return nil, fmt.Errorf("store: referrer not found: %w", err)
+	}
+	return &ref, nil
+}
+
+// GetReferrerByAccount returns the referrer record for an account.
+func (s *PostgresStore) GetReferrerByAccount(accountID string) (*Referrer, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var ref Referrer
+	err := s.pool.QueryRow(ctx,
+		`SELECT account_id, code, created_at FROM referrers WHERE account_id = $1`, accountID,
+	).Scan(&ref.AccountID, &ref.Code, &ref.CreatedAt)
+	if err != nil {
+		return nil, fmt.Errorf("store: referrer not found: %w", err)
+	}
+	return &ref, nil
+}
+
+// RecordReferral records that referredAccountID was referred by referrerCode.
+func (s *PostgresStore) RecordReferral(referrerCode, referredAccountID string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := s.pool.Exec(ctx,
+		`INSERT INTO referrals (referred_account, referrer_code) VALUES ($1, $2)`,
+		referredAccountID, referrerCode,
+	)
+	if err != nil {
+		return fmt.Errorf("store: record referral: %w", err)
+	}
+	return nil
+}
+
+// GetReferrerForAccount returns the referrer code that referred this account.
+func (s *PostgresStore) GetReferrerForAccount(accountID string) (string, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var code string
+	err := s.pool.QueryRow(ctx,
+		`SELECT referrer_code FROM referrals WHERE referred_account = $1`, accountID,
+	).Scan(&code)
+	if err != nil {
+		return "", nil // no referrer is not an error
+	}
+	return code, nil
+}
+
+// GetReferralStats returns referral statistics for a code.
+func (s *PostgresStore) GetReferralStats(code string) (*ReferralStats, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	// Verify code exists
+	var accountID string
+	err := s.pool.QueryRow(ctx,
+		`SELECT account_id FROM referrers WHERE code = $1`, code,
+	).Scan(&accountID)
+	if err != nil {
+		return nil, fmt.Errorf("store: referral code not found: %w", err)
+	}
+
+	// Count referred accounts
+	var totalReferred int
+	_ = s.pool.QueryRow(ctx,
+		`SELECT COUNT(*) FROM referrals WHERE referrer_code = $1`, code,
+	).Scan(&totalReferred)
+
+	// Sum referral rewards from ledger
+	var totalRewards int64
+	_ = s.pool.QueryRow(ctx,
+		`SELECT COALESCE(SUM(amount_micro_usd), 0) FROM ledger_entries
+		 WHERE account_id = $1 AND entry_type = $2`,
+		accountID, string(LedgerReferralReward),
+	).Scan(&totalRewards)
+
+	return &ReferralStats{
+		Code:                 code,
+		TotalReferred:        totalReferred,
+		TotalRewardsMicroUSD: totalRewards,
+	}, nil
+}
