+++ coordinator/store/postgres_users.go
@@ -1,0 +1,162 @@
+package store
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// --- Users (Privy) ---
+
+// CreateUser creates a new user record linked to a Privy identity.
+func (s *PostgresStore) CreateUser(user *User) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := s.pool.Exec(ctx,
+		`INSERT INTO users (account_id, privy_user_id, email, role, platform_fee_percent)
+		 VALUES ($1, $2, $3, $4, $5)`,
+		user.AccountID, user.PrivyUserID, user.Email, user.Role, user.PlatformFeePercent,
+	)
+	if err != nil {
+		return fmt.Errorf("store: create user: %w", err)
+	}
+	return nil
+}
+
+const userSelectColumns = `account_id, privy_user_id, email, role, platform_fee_percent,
+	stripe_account_id, stripe_account_status, stripe_destination_type,
+	stripe_destination_last4, stripe_instant_eligible, created_at`
+
+func scanUser(row interface {
+	Scan(...any) error
+}) (*User, error) {
+	var u User
+	if err := row.Scan(&u.AccountID, &u.PrivyUserID, &u.Email, &u.Role, &u.PlatformFeePercent,
+		&u.StripeAccountID, &u.StripeAccountStatus, &u.StripeDestinationType,
+		&u.StripeDestinationLast4, &u.StripeInstantEligible, &u.CreatedAt); err != nil {
+		return nil, err
+	}
+	return &u, nil
+}
+
+// GetUserByPrivyID returns the user for a Privy DID.
+func (s *PostgresStore) GetUserByPrivyID(privyUserID string) (*User, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	row := s.pool.QueryRow(ctx,
+		`SELECT `+userSelectColumns+` FROM users WHERE privy_user_id = $1`, privyUserID,
+	)
+	u, err := scanUser(row)
+	if err != nil {
+		return nil, fmt.Errorf("store: user not found: %w", err)
+	}
+	return u, nil
+}
+
+// GetUserByAccountID returns the user for an internal account ID.
+func (s *PostgresStore) GetUserByAccountID(accountID string) (*User, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	row := s.pool.QueryRow(ctx,
+		`SELECT `+userSelectColumns+` FROM users WHERE account_id = $1`, accountID,
+	)
+	u, err := scanUser(row)
+	if err != nil {
+		return nil, fmt.Errorf("store: user not found: %w", err)
+	}
+	return u, nil
+}
+
+// SetUserStripeAccount upserts the Stripe Connect fields on a user record.
+func (s *PostgresStore) SetUserStripeAccount(accountID, stripeAccountID, status, destinationType, destinationLast4 string, instantEligible bool) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tag, err := s.pool.Exec(ctx,
+		`UPDATE users SET
+			stripe_account_id = $2,
+			stripe_account_status = $3,
+			stripe_destination_type = $4,
+			stripe_destination_last4 = $5,
+			stripe_instant_eligible = $6
+		 WHERE account_id = $1`,
+		accountID, stripeAccountID, status, destinationType, destinationLast4, instantEligible,
+	)
+	if err != nil {
+		return fmt.Errorf("store: set stripe account: %w", err)
+	}
+	if tag.RowsAffected() == 0 {
+		return fmt.Errorf("user with account ID %q not found", accountID)
+	}
+	return nil
+}
+
+// GetUserByStripeAccount finds a user by their Stripe connected account ID.
+func (s *PostgresStore) GetUserByStripeAccount(stripeAccountID string) (*User, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	row := s.pool.QueryRow(ctx,
+		`SELECT `+userSelectColumns+` FROM users WHERE stripe_account_id = $1`, stripeAccountID,
+	)
+	u, err := scanUser(row)
+	if err != nil {
+		return nil, fmt.Errorf("store: user with Stripe account %q not found: %w", stripeAccountID, err)
+	}
+	return u, nil
+}
+
+// SetUserRole sets the account role on a user record.
+func (s *PostgresStore) SetUserRole(accountID, role string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tag, err := s.pool.Exec(ctx,
+		`UPDATE users SET role = $2 WHERE account_id = $1`,
+		accountID, role,
+	)
+	if err != nil {
+		return fmt.Errorf("store: set user role: %w", err)
+	}
+	if tag.RowsAffected() == 0 {
+		return fmt.Errorf("user with account ID %q not found", accountID)
+	}
+	return nil
+}
+
+// SetUserPlatformFeePercent sets (or clears, when nil) the per-account platform
+// fee override.
+func (s *PostgresStore) SetUserPlatformFeePercent(accountID string, feePercent *int64) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tag, err := s.pool.Exec(ctx,
+		`UPDATE users SET platform_fee_percent = $2 WHERE account_id = $1`,
+		accountID, feePercent,
+	)
+	if err != nil {
+		return fmt.Errorf("store: set user platform fee: %w", err)
+	}
+	if tag.RowsAffected() == 0 {
+		return fmt.Errorf("user with account ID %q not found", accountID)
+	}
+	return nil
+}
+
+// GetUserByEmail returns the user for an email address.
+func (s *PostgresStore) GetUserByEmail(email string) (*User, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	row := s.pool.QueryRow(ctx,
+		`SELECT `+userSelectColumns+` FROM users WHERE LOWER(email) = LOWER($1)`, email,
+	)
+	u, err := scanUser(row)
+	if err != nil {
+		return nil, fmt.Errorf("user with email %q not found", email)
+	}
+	return u, nil
+}
